Create a simplified report of a model's summary

Create a simplified report of a model's summary

report_model(model_object = NULL, response_name = "Score")

Arguments

model_object
A model object

response_name
Name of the response variable. Defaults to "Score".

Value

A data.frame object showing a simple model report that includes the effect of each predictor variable on the response.

Examples

models<-fit_models(df=yields,yname=c("height","yield"),xname="weight",
modeltype=c("lm", "glm"))
report_model(models[[2]][[1]])
#>              Type   Estimate       P_Value Exp_Estimate  Effect
#> 1 Estimated Score  0.5661263 1.446829e-233    1.7614306    1.76
#> 2          weight -0.2173539  1.470866e-08    0.8046452 -19.54%