A convenient selector gadget

A convenient selector gadget

select_col(df, ...)

Arguments

df
The data set from which to select a column

...
columns to select, no quotes

Value

Returns a dataframe with selected columns

Details

A friendly way to select a column or several columns. Mainly for non-pipe usage It is recommended to use known select functions to do pipe manipulations. Otherwise convert to tibble

Examples

select_col(yields,height,weight,normal)

#>           height      weight normal
#> 1     0.28490898 0.134423125    Yes
#> 2     0.24278264 0.374849713     No
#> 3     0.25794322 0.471348277    Yes
#> 4     0.51756041 0.501435921     No
#> 5     0.40260229 0.471717555    Yes
#> 6     0.97898856 0.041919373     No
#> 7     0.37630519 0.270802170    Yes
#> 8     0.15030532 0.582584340     No
#> 9     0.65094469 0.436970887    Yes
#> 10    0.44144997 0.010409170     No
#> 11    0.70782421 0.096153317    Yes
#> 12    0.29768528 0.355558874     No
#> 13    0.57683552 0.466442367    Yes
#> 14    0.28672801 0.056129793     No
#> 15    0.48073405 0.099503653    Yes
#> 16    0.76800428 0.105509859     No
#> 17    0.52065851 0.451008492    Yes
#> 18    0.29753286 0.542230731     No
#> 19    0.60576005 0.217237120    Yes
#> 20    0.55740152 0.097001268     No
#> 21    0.40078678 0.266465657    Yes
#> 22    0.19936941 0.514106114     No
#> 23    0.42644088 0.112576448    Yes
#> 24    0.41888542 0.552212434     No
#> 25    0.49520211 0.237739326    Yes
#> 26    0.34117591 0.011898441     No
#> 27    0.50977338 0.202322198    Yes
#> 28    0.56275339 0.117395468     No
#> 29    0.66183151 0.115626798    Yes
#> 30    0.31800476 0.013107151     No
#> 31    0.41687145 0.475938666    Yes
#> 32    0.31804226 0.297385642     No
#> 33    0.37867445 0.120684234    Yes
#> 34    0.76001789 0.417668229     No
#> 35    0.68785723 0.386476372    Yes
#> 36    0.74286368 0.507571179     No
#> 37    0.68987626 0.275461530    Yes
#> 38    0.60253223 0.136451178     No
#> 39    0.13787241 0.411520463    Yes
#> 40    0.41399654 0.595210123     No
#> 41    0.66618660 0.261146491    Yes
#> 42    0.16747143 0.472149954     No
#> 43    0.36363378 0.365655581    Yes
#> 44    0.72945052 0.389157147     No
#> 45    0.40211495 0.135884730    Yes
#> 46    0.87177815 0.006516951     No
#> 47    0.49172979 0.042850908    Yes
#> 48    0.66353534 0.012354141     No
#> 49    0.54816316 0.186607659    Yes
#> 50    0.43927018 0.195573613     No
#> 51    0.58766825 0.539660962    Yes
#> 52    0.19372644 0.461853811     No
#> 53    0.36424137 0.071836387    Yes
#> 54    0.60268532 0.217212784     No
#> 55    0.50814950 0.160573053    Yes
#> 56    0.57403088 0.535286790     No
#> 57    0.80891580 0.063551108    Yes
#> 58    0.66080555 0.207301304     No
#> 59    0.35262665 0.131934818    Yes
#> 60    0.54119871 0.327473600     No
#> 61    0.52045849 0.126929427    Yes
#> 62    0.18797580 0.593378821     No
#> 63    0.68585242 0.372396663    Yes
#> 64    0.99616135 0.318422948     No
#> 65    0.55802169 0.195180455    Yes
#> 66    0.61090853 0.232337822     No
#> 67    0.25017721 0.425583519    Yes
#> 68    0.39202214 0.207503032     No
#> 69    0.56014276 0.372323025    Yes
#> 70    0.40964296 0.454286314     No
#> 71    0.43422640 0.132933513    Yes
#> 72    0.66875441 0.011974596     No
#> 73    0.40035979 0.407187511    Yes
#> 74    0.67225415 0.378223932     No
#> 75    0.45543231 0.312028499    Yes
#> 76    0.76741789 0.200362143     No
#> 77    0.43224602 0.240926166    Yes
#> 78    0.08529002 0.579631291     No
#> 79    0.41382512 0.465168390    Yes
#> 80    0.97066447 0.328370631     No
#> 81    0.55102534 0.342261183    Yes
#> 82    0.48650110 0.441905645     No
#> 83    0.66342761 0.555620749    Yes
#> 84    0.73913575 0.309384938     No
#> 85   -0.15102802 0.327702073    Yes
#> 86    0.77917734 0.450656877     No
#> 87    0.61397708 0.249915379    Yes
#> 88    0.82961103 0.370494035     No
#> 89    0.61947372 0.277727210    Yes
#> 90    0.55337617 0.095705962     No
#> 91    0.26634638 0.014462903    Yes
#> 92    0.36279149 0.459377337     No
#> 93    0.77610074 0.499797150    Yes
#> 94    0.97343609 0.036536516     No
#> 95    0.61348425 0.285210663    Yes
#> 96    0.05957072 0.053203518     No
#> 97    0.55377182 0.473441250    Yes
#> 98    0.35417790 0.545332176     No
#> 99    0.54177183 0.592745642    Yes
#> 100   0.17425675 0.176117167     No
#> 101   0.28490898 0.134423125    Yes
#> 102   0.24278264 0.374849713     No
#> 103   0.25794322 0.471348277    Yes
#> 104   0.51756041 0.501435921     No
#> 105   0.40260229 0.471717555    Yes
#> 106   0.97898856 0.041919373     No
#> 107   0.37630519 0.270802170    Yes
#> 108   0.15030532 0.582584340     No
#> 109   0.65094469 0.436970887    Yes
#> 110   0.44144997 0.010409170     No
#> 111   0.70782421 0.096153317    Yes
#> 112   0.29768528 0.355558874     No
#> 113   0.57683552 0.466442367    Yes
#> 114   0.28672801 0.056129793     No
#> 115   0.48073405 0.099503653    Yes
#> 116   0.76800428 0.105509859     No
#> 117   0.52065851 0.451008492    Yes
#> 118   0.29753286 0.542230731     No
#> 119   0.60576005 0.217237120    Yes
#> 120   0.55740152 0.097001268     No
#> 121   0.40078678 0.266465657    Yes
#> 122   0.19936941 0.514106114     No
#> 123   0.42644088 0.112576448    Yes
#> 124   0.41888542 0.552212434     No
#> 125   0.49520211 0.237739326    Yes
#> 126   0.34117591 0.011898441     No
#> 127   0.50977338 0.202322198    Yes
#> 128   0.56275339 0.117395468     No
#> 129   0.66183151 0.115626798    Yes
#> 130   0.31800476 0.013107151     No
#> 131   0.41687145 0.475938666    Yes
#> 132   0.31804226 0.297385642     No
#> 133   0.37867445 0.120684234    Yes
#> 134   0.76001789 0.417668229     No
#> 135   0.68785723 0.386476372    Yes
#> 136   0.74286368 0.507571179     No
#> 137   0.68987626 0.275461530    Yes
#> 138   0.60253223 0.136451178     No
#> 139   0.13787241 0.411520463    Yes
#> 140   0.41399654 0.595210123     No
#> 141   0.66618660 0.261146491    Yes
#> 142   0.16747143 0.472149954     No
#> 143   0.36363378 0.365655581    Yes
#> 144   0.72945052 0.389157147     No
#> 145   0.40211495 0.135884730    Yes
#> 146   0.87177815 0.006516951     No
#> 147   0.49172979 0.042850908    Yes
#> 148   0.66353534 0.012354141     No
#> 149   0.54816316 0.186607659    Yes
#> 150   0.43927018 0.195573613     No
#> 151   0.58766825 0.539660962    Yes
#> 152   0.19372644 0.461853811     No
#> 153   0.36424137 0.071836387    Yes
#> 154   0.60268532 0.217212784     No
#> 155   0.50814950 0.160573053    Yes
#> 156   0.57403088 0.535286790     No
#> 157   0.80891580 0.063551108    Yes
#> 158   0.66080555 0.207301304     No
#> 159   0.35262665 0.131934818    Yes
#> 160   0.54119871 0.327473600     No
#> 161   0.52045849 0.126929427    Yes
#> 162   0.18797580 0.593378821     No
#> 163   0.68585242 0.372396663    Yes
#> 164   0.99616135 0.318422948     No
#> 165   0.55802169 0.195180455    Yes
#> 166   0.61090853 0.232337822     No
#> 167   0.25017721 0.425583519    Yes
#> 168   0.39202214 0.207503032     No
#> 169   0.56014276 0.372323025    Yes
#> 170   0.40964296 0.454286314     No
#> 171   0.43422640 0.132933513    Yes
#> 172   0.66875441 0.011974596     No
#> 173   0.40035979 0.407187511    Yes
#> 174   0.67225415 0.378223932     No
#> 175   0.45543231 0.312028499    Yes
#> 176   0.76741789 0.200362143     No
#> 177   0.43224602 0.240926166    Yes
#> 178   0.08529002 0.579631291     No
#> 179   0.41382512 0.465168390    Yes
#> 180   0.97066447 0.328370631     No
#> 181   0.55102534 0.342261183    Yes
#> 182   0.48650110 0.441905645     No
#> 183   0.66342761 0.555620749    Yes
#> 184   0.73913575 0.309384938     No
#> 185  -0.15102802 0.327702073    Yes
#> 186   0.77917734 0.450656877     No
#> 187   0.61397708 0.249915379    Yes
#> 188   0.82961103 0.370494035     No
#> 189   0.61947372 0.277727210    Yes
#> 190   0.55337617 0.095705962     No
#> 191   0.26634638 0.014462903    Yes
#> 192   0.36279149 0.459377337     No
#> 193   0.77610074 0.499797150    Yes
#> 194   0.97343609 0.036536516     No
#> 195   0.61348425 0.285210663    Yes
#> 196   0.05957072 0.053203518     No
#> 197   0.55377182 0.473441250    Yes
#> 198   0.35417790 0.545332176     No
#> 199   0.54177183 0.592745642    Yes
#> 200   0.17425675 0.176117167     No
#> 201   0.28490898 0.134423125    Yes
#> 202   0.24278264 0.374849713     No
#> 203   0.25794322 0.471348277    Yes
#> 204   0.51756041 0.501435921     No
#> 205   0.40260229 0.471717555    Yes
#> 206   0.97898856 0.041919373     No
#> 207   0.37630519 0.270802170    Yes
#> 208   0.15030532 0.582584340     No
#> 209   0.65094469 0.436970887    Yes
#> 210   0.44144997 0.010409170     No
#> 211   0.70782421 0.096153317    Yes
#> 212   0.29768528 0.355558874     No
#> 213   0.57683552 0.466442367    Yes
#> 214   0.28672801 0.056129793     No
#> 215   0.48073405 0.099503653    Yes
#> 216   0.76800428 0.105509859     No
#> 217   0.52065851 0.451008492    Yes
#> 218   0.29753286 0.542230731     No
#> 219   0.60576005 0.217237120    Yes
#> 220   0.55740152 0.097001268     No
#> 221   0.40078678 0.266465657    Yes
#> 222   0.19936941 0.514106114     No
#> 223   0.42644088 0.112576448    Yes
#> 224   0.41888542 0.552212434     No
#> 225   0.49520211 0.237739326    Yes
#> 226   0.34117591 0.011898441     No
#> 227   0.50977338 0.202322198    Yes
#> 228   0.56275339 0.117395468     No
#> 229   0.66183151 0.115626798    Yes
#> 230   0.31800476 0.013107151     No
#> 231   0.41687145 0.475938666    Yes
#> 232   0.31804226 0.297385642     No
#> 233   0.37867445 0.120684234    Yes
#> 234   0.76001789 0.417668229     No
#> 235   0.68785723 0.386476372    Yes
#> 236   0.74286368 0.507571179     No
#> 237   0.68987626 0.275461530    Yes
#> 238   0.60253223 0.136451178     No
#> 239   0.13787241 0.411520463    Yes
#> 240   0.41399654 0.595210123     No
#> 241   0.66618660 0.261146491    Yes
#> 242   0.16747143 0.472149954     No
#> 243   0.36363378 0.365655581    Yes
#> 244   0.72945052 0.389157147     No
#> 245   0.40211495 0.135884730    Yes
#> 246   0.87177815 0.006516951     No
#> 247   0.49172979 0.042850908    Yes
#> 248   0.66353534 0.012354141     No
#> 249   0.54816316 0.186607659    Yes
#> 250   0.43927018 0.195573613     No
#> 251   0.58766825 0.539660962    Yes
#> 252   0.19372644 0.461853811     No
#> 253   0.36424137 0.071836387    Yes
#> 254   0.60268532 0.217212784     No
#> 255   0.50814950 0.160573053    Yes
#> 256   0.57403088 0.535286790     No
#> 257   0.80891580 0.063551108    Yes
#> 258   0.66080555 0.207301304     No
#> 259   0.35262665 0.131934818    Yes
#> 260   0.54119871 0.327473600     No
#> 261   0.52045849 0.126929427    Yes
#> 262   0.18797580 0.593378821     No
#> 263   0.68585242 0.372396663    Yes
#> 264   0.99616135 0.318422948     No
#> 265   0.55802169 0.195180455    Yes
#> 266   0.61090853 0.232337822     No
#> 267   0.25017721 0.425583519    Yes
#> 268   0.39202214 0.207503032     No
#> 269   0.56014276 0.372323025    Yes
#> 270   0.40964296 0.454286314     No
#> 271   0.43422640 0.132933513    Yes
#> 272   0.66875441 0.011974596     No
#> 273   0.40035979 0.407187511    Yes
#> 274   0.67225415 0.378223932     No
#> 275   0.45543231 0.312028499    Yes
#> 276   0.76741789 0.200362143     No
#> 277   0.43224602 0.240926166    Yes
#> 278   0.08529002 0.579631291     No
#> 279   0.41382512 0.465168390    Yes
#> 280   0.97066447 0.328370631     No
#> 281   0.55102534 0.342261183    Yes
#> 282   0.48650110 0.441905645     No
#> 283   0.66342761 0.555620749    Yes
#> 284   0.73913575 0.309384938     No
#> 285  -0.15102802 0.327702073    Yes
#> 286   0.77917734 0.450656877     No
#> 287   0.61397708 0.249915379    Yes
#> 288   0.82961103 0.370494035     No
#> 289   0.61947372 0.277727210    Yes
#> 290   0.55337617 0.095705962     No
#> 291   0.26634638 0.014462903    Yes
#> 292   0.36279149 0.459377337     No
#> 293   0.77610074 0.499797150    Yes
#> 294   0.97343609 0.036536516     No
#> 295   0.61348425 0.285210663    Yes
#> 296   0.05957072 0.053203518     No
#> 297   0.55377182 0.473441250    Yes
#> 298   0.35417790 0.545332176     No
#> 299   0.54177183 0.592745642    Yes
#> 300   0.17425675 0.176117167     No
#> 301   0.28490898 0.134423125    Yes
#> 302   0.24278264 0.374849713     No
#> 303   0.25794322 0.471348277    Yes
#> 304   0.51756041 0.501435921     No
#> 305   0.40260229 0.471717555    Yes
#> 306   0.97898856 0.041919373     No
#> 307   0.37630519 0.270802170    Yes
#> 308   0.15030532 0.582584340     No
#> 309   0.65094469 0.436970887    Yes
#> 310   0.44144997 0.010409170     No
#> 311   0.70782421 0.096153317    Yes
#> 312   0.29768528 0.355558874     No
#> 313   0.57683552 0.466442367    Yes
#> 314   0.28672801 0.056129793     No
#> 315   0.48073405 0.099503653    Yes
#> 316   0.76800428 0.105509859     No
#> 317   0.52065851 0.451008492    Yes
#> 318   0.29753286 0.542230731     No
#> 319   0.60576005 0.217237120    Yes
#> 320   0.55740152 0.097001268     No
#> 321   0.40078678 0.266465657    Yes
#> 322   0.19936941 0.514106114     No
#> 323   0.42644088 0.112576448    Yes
#> 324   0.41888542 0.552212434     No
#> 325   0.49520211 0.237739326    Yes
#> 326   0.34117591 0.011898441     No
#> 327   0.50977338 0.202322198    Yes
#> 328   0.56275339 0.117395468     No
#> 329   0.66183151 0.115626798    Yes
#> 330   0.31800476 0.013107151     No
#> 331   0.41687145 0.475938666    Yes
#> 332   0.31804226 0.297385642     No
#> 333   0.37867445 0.120684234    Yes
#> 334   0.76001789 0.417668229     No
#> 335   0.68785723 0.386476372    Yes
#> 336   0.74286368 0.507571179     No
#> 337   0.68987626 0.275461530    Yes
#> 338   0.60253223 0.136451178     No
#> 339   0.13787241 0.411520463    Yes
#> 340   0.41399654 0.595210123     No
#> 341   0.66618660 0.261146491    Yes
#> 342   0.16747143 0.472149954     No
#> 343   0.36363378 0.365655581    Yes
#> 344   0.72945052 0.389157147     No
#> 345   0.40211495 0.135884730    Yes
#> 346   0.87177815 0.006516951     No
#> 347   0.49172979 0.042850908    Yes
#> 348   0.66353534 0.012354141     No
#> 349   0.54816316 0.186607659    Yes
#> 350   0.43927018 0.195573613     No
#> 351   0.58766825 0.539660962    Yes
#> 352   0.19372644 0.461853811     No
#> 353   0.36424137 0.071836387    Yes
#> 354   0.60268532 0.217212784     No
#> 355   0.50814950 0.160573053    Yes
#> 356   0.57403088 0.535286790     No
#> 357   0.80891580 0.063551108    Yes
#> 358   0.66080555 0.207301304     No
#> 359   0.35262665 0.131934818    Yes
#> 360   0.54119871 0.327473600     No
#> 361   0.52045849 0.126929427    Yes
#> 362   0.18797580 0.593378821     No
#> 363   0.68585242 0.372396663    Yes
#> 364   0.99616135 0.318422948     No
#> 365   0.55802169 0.195180455    Yes
#> 366   0.61090853 0.232337822     No
#> 367   0.25017721 0.425583519    Yes
#> 368   0.39202214 0.207503032     No
#> 369   0.56014276 0.372323025    Yes
#> 370   0.40964296 0.454286314     No
#> 371   0.43422640 0.132933513    Yes
#> 372   0.66875441 0.011974596     No
#> 373   0.40035979 0.407187511    Yes
#> 374   0.67225415 0.378223932     No
#> 375   0.45543231 0.312028499    Yes
#> 376   0.76741789 0.200362143     No
#> 377   0.43224602 0.240926166    Yes
#> 378   0.08529002 0.579631291     No
#> 379   0.41382512 0.465168390    Yes
#> 380   0.97066447 0.328370631     No
#> 381   0.55102534 0.342261183    Yes
#> 382   0.48650110 0.441905645     No
#> 383   0.66342761 0.555620749    Yes
#> 384   0.73913575 0.309384938     No
#> 385  -0.15102802 0.327702073    Yes
#> 386   0.77917734 0.450656877     No
#> 387   0.61397708 0.249915379    Yes
#> 388   0.82961103 0.370494035     No
#> 389   0.61947372 0.277727210    Yes
#> 390   0.55337617 0.095705962     No
#> 391   0.26634638 0.014462903    Yes
#> 392   0.36279149 0.459377337     No
#> 393   0.77610074 0.499797150    Yes
#> 394   0.97343609 0.036536516     No
#> 395   0.61348425 0.285210663    Yes
#> 396   0.05957072 0.053203518     No
#> 397   0.55377182 0.473441250    Yes
#> 398   0.35417790 0.545332176     No
#> 399   0.54177183 0.592745642    Yes
#> 400   0.17425675 0.176117167     No
#> 401   0.28490898 0.134423125    Yes
#> 402   0.24278264 0.374849713     No
#> 403   0.25794322 0.471348277    Yes
#> 404   0.51756041 0.501435921     No
#> 405   0.40260229 0.471717555    Yes
#> 406   0.97898856 0.041919373     No
#> 407   0.37630519 0.270802170    Yes
#> 408   0.15030532 0.582584340     No
#> 409   0.65094469 0.436970887    Yes
#> 410   0.44144997 0.010409170     No
#> 411   0.70782421 0.096153317    Yes
#> 412   0.29768528 0.355558874     No
#> 413   0.57683552 0.466442367    Yes
#> 414   0.28672801 0.056129793     No
#> 415   0.48073405 0.099503653    Yes
#> 416   0.76800428 0.105509859     No
#> 417   0.52065851 0.451008492    Yes
#> 418   0.29753286 0.542230731     No
#> 419   0.60576005 0.217237120    Yes
#> 420   0.55740152 0.097001268     No
#> 421   0.40078678 0.266465657    Yes
#> 422   0.19936941 0.514106114     No
#> 423   0.42644088 0.112576448    Yes
#> 424   0.41888542 0.552212434     No
#> 425   0.49520211 0.237739326    Yes
#> 426   0.34117591 0.011898441     No
#> 427   0.50977338 0.202322198    Yes
#> 428   0.56275339 0.117395468     No
#> 429   0.66183151 0.115626798    Yes
#> 430   0.31800476 0.013107151     No
#> 431   0.41687145 0.475938666    Yes
#> 432   0.31804226 0.297385642     No
#> 433   0.37867445 0.120684234    Yes
#> 434   0.76001789 0.417668229     No
#> 435   0.68785723 0.386476372    Yes
#> 436   0.74286368 0.507571179     No
#> 437   0.68987626 0.275461530    Yes
#> 438   0.60253223 0.136451178     No
#> 439   0.13787241 0.411520463    Yes
#> 440   0.41399654 0.595210123     No
#> 441   0.66618660 0.261146491    Yes
#> 442   0.16747143 0.472149954     No
#> 443   0.36363378 0.365655581    Yes
#> 444   0.72945052 0.389157147     No
#> 445   0.40211495 0.135884730    Yes
#> 446   0.87177815 0.006516951     No
#> 447   0.49172979 0.042850908    Yes
#> 448   0.66353534 0.012354141     No
#> 449   0.54816316 0.186607659    Yes
#> 450   0.43927018 0.195573613     No
#> 451   0.58766825 0.539660962    Yes
#> 452   0.19372644 0.461853811     No
#> 453   0.36424137 0.071836387    Yes
#> 454   0.60268532 0.217212784     No
#> 455   0.50814950 0.160573053    Yes
#> 456   0.57403088 0.535286790     No
#> 457   0.80891580 0.063551108    Yes
#> 458   0.66080555 0.207301304     No
#> 459   0.35262665 0.131934818    Yes
#> 460   0.54119871 0.327473600     No
#> 461   0.52045849 0.126929427    Yes
#> 462   0.18797580 0.593378821     No
#> 463   0.68585242 0.372396663    Yes
#> 464   0.99616135 0.318422948     No
#> 465   0.55802169 0.195180455    Yes
#> 466   0.61090853 0.232337822     No
#> 467   0.25017721 0.425583519    Yes
#> 468   0.39202214 0.207503032     No
#> 469   0.56014276 0.372323025    Yes
#> 470   0.40964296 0.454286314     No
#> 471   0.43422640 0.132933513    Yes
#> 472   0.66875441 0.011974596     No
#> 473   0.40035979 0.407187511    Yes
#> 474   0.67225415 0.378223932     No
#> 475   0.45543231 0.312028499    Yes
#> 476   0.76741789 0.200362143     No
#> 477   0.43224602 0.240926166    Yes
#> 478   0.08529002 0.579631291     No
#> 479   0.41382512 0.465168390    Yes
#> 480   0.97066447 0.328370631     No
#> 481   0.55102534 0.342261183    Yes
#> 482   0.48650110 0.441905645     No
#> 483   0.66342761 0.555620749    Yes
#> 484   0.73913575 0.309384938     No
#> 485  -0.15102802 0.327702073    Yes
#> 486   0.77917734 0.450656877     No
#> 487   0.61397708 0.249915379    Yes
#> 488   0.82961103 0.370494035     No
#> 489   0.61947372 0.277727210    Yes
#> 490   0.55337617 0.095705962     No
#> 491   0.26634638 0.014462903    Yes
#> 492   0.36279149 0.459377337     No
#> 493   0.77610074 0.499797150    Yes
#> 494   0.97343609 0.036536516     No
#> 495   0.61348425 0.285210663    Yes
#> 496   0.05957072 0.053203518     No
#> 497   0.55377182 0.473441250    Yes
#> 498   0.35417790 0.545332176     No
#> 499   0.54177183 0.592745642    Yes
#> 500   0.17425675 0.176117167     No
#> 501   0.28490898 0.134423125    Yes
#> 502   0.24278264 0.374849713     No
#> 503   0.25794322 0.471348277    Yes
#> 504   0.51756041 0.501435921     No
#> 505   0.40260229 0.471717555    Yes
#> 506   0.97898856 0.041919373     No
#> 507   0.37630519 0.270802170    Yes
#> 508   0.15030532 0.582584340     No
#> 509   0.65094469 0.436970887    Yes
#> 510   0.44144997 0.010409170     No
#> 511   0.70782421 0.096153317    Yes
#> 512   0.29768528 0.355558874     No
#> 513   0.57683552 0.466442367    Yes
#> 514   0.28672801 0.056129793     No
#> 515   0.48073405 0.099503653    Yes
#> 516   0.76800428 0.105509859     No
#> 517   0.52065851 0.451008492    Yes
#> 518   0.29753286 0.542230731     No
#> 519   0.60576005 0.217237120    Yes
#> 520   0.55740152 0.097001268     No
#> 521   0.40078678 0.266465657    Yes
#> 522   0.19936941 0.514106114     No
#> 523   0.42644088 0.112576448    Yes
#> 524   0.41888542 0.552212434     No
#> 525   0.49520211 0.237739326    Yes
#> 526   0.34117591 0.011898441     No
#> 527   0.50977338 0.202322198    Yes
#> 528   0.56275339 0.117395468     No
#> 529   0.66183151 0.115626798    Yes
#> 530   0.31800476 0.013107151     No
#> 531   0.41687145 0.475938666    Yes
#> 532   0.31804226 0.297385642     No
#> 533   0.37867445 0.120684234    Yes
#> 534   0.76001789 0.417668229     No
#> 535   0.68785723 0.386476372    Yes
#> 536   0.74286368 0.507571179     No
#> 537   0.68987626 0.275461530    Yes
#> 538   0.60253223 0.136451178     No
#> 539   0.13787241 0.411520463    Yes
#> 540   0.41399654 0.595210123     No
#> 541   0.66618660 0.261146491    Yes
#> 542   0.16747143 0.472149954     No
#> 543   0.36363378 0.365655581    Yes
#> 544   0.72945052 0.389157147     No
#> 545   0.40211495 0.135884730    Yes
#> 546   0.87177815 0.006516951     No
#> 547   0.49172979 0.042850908    Yes
#> 548   0.66353534 0.012354141     No
#> 549   0.54816316 0.186607659    Yes
#> 550   0.43927018 0.195573613     No
#> 551   0.58766825 0.539660962    Yes
#> 552   0.19372644 0.461853811     No
#> 553   0.36424137 0.071836387    Yes
#> 554   0.60268532 0.217212784     No
#> 555   0.50814950 0.160573053    Yes
#> 556   0.57403088 0.535286790     No
#> 557   0.80891580 0.063551108    Yes
#> 558   0.66080555 0.207301304     No
#> 559   0.35262665 0.131934818    Yes
#> 560   0.54119871 0.327473600     No
#> 561   0.52045849 0.126929427    Yes
#> 562   0.18797580 0.593378821     No
#> 563   0.68585242 0.372396663    Yes
#> 564   0.99616135 0.318422948     No
#> 565   0.55802169 0.195180455    Yes
#> 566   0.61090853 0.232337822     No
#> 567   0.25017721 0.425583519    Yes
#> 568   0.39202214 0.207503032     No
#> 569   0.56014276 0.372323025    Yes
#> 570   0.40964296 0.454286314     No
#> 571   0.43422640 0.132933513    Yes
#> 572   0.66875441 0.011974596     No
#> 573   0.40035979 0.407187511    Yes
#> 574   0.67225415 0.378223932     No
#> 575   0.45543231 0.312028499    Yes
#> 576   0.76741789 0.200362143     No
#> 577   0.43224602 0.240926166    Yes
#> 578   0.08529002 0.579631291     No
#> 579   0.41382512 0.465168390    Yes
#> 580   0.97066447 0.328370631     No
#> 581   0.55102534 0.342261183    Yes
#> 582   0.48650110 0.441905645     No
#> 583   0.66342761 0.555620749    Yes
#> 584   0.73913575 0.309384938     No
#> 585  -0.15102802 0.327702073    Yes
#> 586   0.77917734 0.450656877     No
#> 587   0.61397708 0.249915379    Yes
#> 588   0.82961103 0.370494035     No
#> 589   0.61947372 0.277727210    Yes
#> 590   0.55337617 0.095705962     No
#> 591   0.26634638 0.014462903    Yes
#> 592   0.36279149 0.459377337     No
#> 593   0.77610074 0.499797150    Yes
#> 594   0.97343609 0.036536516     No
#> 595   0.61348425 0.285210663    Yes
#> 596   0.05957072 0.053203518     No
#> 597   0.55377182 0.473441250    Yes
#> 598   0.35417790 0.545332176     No
#> 599   0.54177183 0.592745642    Yes
#> 600   0.17425675 0.176117167     No
#> 601   0.28490898 0.134423125    Yes
#> 602   0.24278264 0.374849713     No
#> 603   0.25794322 0.471348277    Yes
#> 604   0.51756041 0.501435921     No
#> 605   0.40260229 0.471717555    Yes
#> 606   0.97898856 0.041919373     No
#> 607   0.37630519 0.270802170    Yes
#> 608   0.15030532 0.582584340     No
#> 609   0.65094469 0.436970887    Yes
#> 610   0.44144997 0.010409170     No
#> 611   0.70782421 0.096153317    Yes
#> 612   0.29768528 0.355558874     No
#> 613   0.57683552 0.466442367    Yes
#> 614   0.28672801 0.056129793     No
#> 615   0.48073405 0.099503653    Yes
#> 616   0.76800428 0.105509859     No
#> 617   0.52065851 0.451008492    Yes
#> 618   0.29753286 0.542230731     No
#> 619   0.60576005 0.217237120    Yes
#> 620   0.55740152 0.097001268     No
#> 621   0.40078678 0.266465657    Yes
#> 622   0.19936941 0.514106114     No
#> 623   0.42644088 0.112576448    Yes
#> 624   0.41888542 0.552212434     No
#> 625   0.49520211 0.237739326    Yes
#> 626   0.34117591 0.011898441     No
#> 627   0.50977338 0.202322198    Yes
#> 628   0.56275339 0.117395468     No
#> 629   0.66183151 0.115626798    Yes
#> 630   0.31800476 0.013107151     No
#> 631   0.41687145 0.475938666    Yes
#> 632   0.31804226 0.297385642     No
#> 633   0.37867445 0.120684234    Yes
#> 634   0.76001789 0.417668229     No
#> 635   0.68785723 0.386476372    Yes
#> 636   0.74286368 0.507571179     No
#> 637   0.68987626 0.275461530    Yes
#> 638   0.60253223 0.136451178     No
#> 639   0.13787241 0.411520463    Yes
#> 640   0.41399654 0.595210123     No
#> 641   0.66618660 0.261146491    Yes
#> 642   0.16747143 0.472149954     No
#> 643   0.36363378 0.365655581    Yes
#> 644   0.72945052 0.389157147     No
#> 645   0.40211495 0.135884730    Yes
#> 646   0.87177815 0.006516951     No
#> 647   0.49172979 0.042850908    Yes
#> 648   0.66353534 0.012354141     No
#> 649   0.54816316 0.186607659    Yes
#> 650   0.43927018 0.195573613     No
#> 651   0.58766825 0.539660962    Yes
#> 652   0.19372644 0.461853811     No
#> 653   0.36424137 0.071836387    Yes
#> 654   0.60268532 0.217212784     No
#> 655   0.50814950 0.160573053    Yes
#> 656   0.57403088 0.535286790     No
#> 657   0.80891580 0.063551108    Yes
#> 658   0.66080555 0.207301304     No
#> 659   0.35262665 0.131934818    Yes
#> 660   0.54119871 0.327473600     No
#> 661   0.52045849 0.126929427    Yes
#> 662   0.18797580 0.593378821     No
#> 663   0.68585242 0.372396663    Yes
#> 664   0.99616135 0.318422948     No
#> 665   0.55802169 0.195180455    Yes
#> 666   0.61090853 0.232337822     No
#> 667   0.25017721 0.425583519    Yes
#> 668   0.39202214 0.207503032     No
#> 669   0.56014276 0.372323025    Yes
#> 670   0.40964296 0.454286314     No
#> 671   0.43422640 0.132933513    Yes
#> 672   0.66875441 0.011974596     No
#> 673   0.40035979 0.407187511    Yes
#> 674   0.67225415 0.378223932     No
#> 675   0.45543231 0.312028499    Yes
#> 676   0.76741789 0.200362143     No
#> 677   0.43224602 0.240926166    Yes
#> 678   0.08529002 0.579631291     No
#> 679   0.41382512 0.465168390    Yes
#> 680   0.97066447 0.328370631     No
#> 681   0.55102534 0.342261183    Yes
#> 682   0.48650110 0.441905645     No
#> 683   0.66342761 0.555620749    Yes
#> 684   0.73913575 0.309384938     No
#> 685  -0.15102802 0.327702073    Yes
#> 686   0.77917734 0.450656877     No
#> 687   0.61397708 0.249915379    Yes
#> 688   0.82961103 0.370494035     No
#> 689   0.61947372 0.277727210    Yes
#> 690   0.55337617 0.095705962     No
#> 691   0.26634638 0.014462903    Yes
#> 692   0.36279149 0.459377337     No
#> 693   0.77610074 0.499797150    Yes
#> 694   0.97343609 0.036536516     No
#> 695   0.61348425 0.285210663    Yes
#> 696   0.05957072 0.053203518     No
#> 697   0.55377182 0.473441250    Yes
#> 698   0.35417790 0.545332176     No
#> 699   0.54177183 0.592745642    Yes
#> 700   0.17425675 0.176117167     No
#> 701   0.28490898 0.134423125    Yes
#> 702   0.24278264 0.374849713     No
#> 703   0.25794322 0.471348277    Yes
#> 704   0.51756041 0.501435921     No
#> 705   0.40260229 0.471717555    Yes
#> 706   0.97898856 0.041919373     No
#> 707   0.37630519 0.270802170    Yes
#> 708   0.15030532 0.582584340     No
#> 709   0.65094469 0.436970887    Yes
#> 710   0.44144997 0.010409170     No
#> 711   0.70782421 0.096153317    Yes
#> 712   0.29768528 0.355558874     No
#> 713   0.57683552 0.466442367    Yes
#> 714   0.28672801 0.056129793     No
#> 715   0.48073405 0.099503653    Yes
#> 716   0.76800428 0.105509859     No
#> 717   0.52065851 0.451008492    Yes
#> 718   0.29753286 0.542230731     No
#> 719   0.60576005 0.217237120    Yes
#> 720   0.55740152 0.097001268     No
#> 721   0.40078678 0.266465657    Yes
#> 722   0.19936941 0.514106114     No
#> 723   0.42644088 0.112576448    Yes
#> 724   0.41888542 0.552212434     No
#> 725   0.49520211 0.237739326    Yes
#> 726   0.34117591 0.011898441     No
#> 727   0.50977338 0.202322198    Yes
#> 728   0.56275339 0.117395468     No
#> 729   0.66183151 0.115626798    Yes
#> 730   0.31800476 0.013107151     No
#> 731   0.41687145 0.475938666    Yes
#> 732   0.31804226 0.297385642     No
#> 733   0.37867445 0.120684234    Yes
#> 734   0.76001789 0.417668229     No
#> 735   0.68785723 0.386476372    Yes
#> 736   0.74286368 0.507571179     No
#> 737   0.68987626 0.275461530    Yes
#> 738   0.60253223 0.136451178     No
#> 739   0.13787241 0.411520463    Yes
#> 740   0.41399654 0.595210123     No
#> 741   0.66618660 0.261146491    Yes
#> 742   0.16747143 0.472149954     No
#> 743   0.36363378 0.365655581    Yes
#> 744   0.72945052 0.389157147     No
#> 745   0.40211495 0.135884730    Yes
#> 746   0.87177815 0.006516951     No
#> 747   0.49172979 0.042850908    Yes
#> 748   0.66353534 0.012354141     No
#> 749   0.54816316 0.186607659    Yes
#> 750   0.43927018 0.195573613     No
#> 751   0.58766825 0.539660962    Yes
#> 752   0.19372644 0.461853811     No
#> 753   0.36424137 0.071836387    Yes
#> 754   0.60268532 0.217212784     No
#> 755   0.50814950 0.160573053    Yes
#> 756   0.57403088 0.535286790     No
#> 757   0.80891580 0.063551108    Yes
#> 758   0.66080555 0.207301304     No
#> 759   0.35262665 0.131934818    Yes
#> 760   0.54119871 0.327473600     No
#> 761   0.52045849 0.126929427    Yes
#> 762   0.18797580 0.593378821     No
#> 763   0.68585242 0.372396663    Yes
#> 764   0.99616135 0.318422948     No
#> 765   0.55802169 0.195180455    Yes
#> 766   0.61090853 0.232337822     No
#> 767   0.25017721 0.425583519    Yes
#> 768   0.39202214 0.207503032     No
#> 769   0.56014276 0.372323025    Yes
#> 770   0.40964296 0.454286314     No
#> 771   0.43422640 0.132933513    Yes
#> 772   0.66875441 0.011974596     No
#> 773   0.40035979 0.407187511    Yes
#> 774   0.67225415 0.378223932     No
#> 775   0.45543231 0.312028499    Yes
#> 776   0.76741789 0.200362143     No
#> 777   0.43224602 0.240926166    Yes
#> 778   0.08529002 0.579631291     No
#> 779   0.41382512 0.465168390    Yes
#> 780   0.97066447 0.328370631     No
#> 781   0.55102534 0.342261183    Yes
#> 782   0.48650110 0.441905645     No
#> 783   0.66342761 0.555620749    Yes
#> 784   0.73913575 0.309384938     No
#> 785  -0.15102802 0.327702073    Yes
#> 786   0.77917734 0.450656877     No
#> 787   0.61397708 0.249915379    Yes
#> 788   0.82961103 0.370494035     No
#> 789   0.61947372 0.277727210    Yes
#> 790   0.55337617 0.095705962     No
#> 791   0.26634638 0.014462903    Yes
#> 792   0.36279149 0.459377337     No
#> 793   0.77610074 0.499797150    Yes
#> 794   0.97343609 0.036536516     No
#> 795   0.61348425 0.285210663    Yes
#> 796   0.05957072 0.053203518     No
#> 797   0.55377182 0.473441250    Yes
#> 798   0.35417790 0.545332176     No
#> 799   0.54177183 0.592745642    Yes
#> 800   0.17425675 0.176117167     No
#> 801   0.28490898 0.134423125    Yes
#> 802   0.24278264 0.374849713     No
#> 803   0.25794322 0.471348277    Yes
#> 804   0.51756041 0.501435921     No
#> 805   0.40260229 0.471717555    Yes
#> 806   0.97898856 0.041919373     No
#> 807   0.37630519 0.270802170    Yes
#> 808   0.15030532 0.582584340     No
#> 809   0.65094469 0.436970887    Yes
#> 810   0.44144997 0.010409170     No
#> 811   0.70782421 0.096153317    Yes
#> 812   0.29768528 0.355558874     No
#> 813   0.57683552 0.466442367    Yes
#> 814   0.28672801 0.056129793     No
#> 815   0.48073405 0.099503653    Yes
#> 816   0.76800428 0.105509859     No
#> 817   0.52065851 0.451008492    Yes
#> 818   0.29753286 0.542230731     No
#> 819   0.60576005 0.217237120    Yes
#> 820   0.55740152 0.097001268     No
#> 821   0.40078678 0.266465657    Yes
#> 822   0.19936941 0.514106114     No
#> 823   0.42644088 0.112576448    Yes
#> 824   0.41888542 0.552212434     No
#> 825   0.49520211 0.237739326    Yes
#> 826   0.34117591 0.011898441     No
#> 827   0.50977338 0.202322198    Yes
#> 828   0.56275339 0.117395468     No
#> 829   0.66183151 0.115626798    Yes
#> 830   0.31800476 0.013107151     No
#> 831   0.41687145 0.475938666    Yes
#> 832   0.31804226 0.297385642     No
#> 833   0.37867445 0.120684234    Yes
#> 834   0.76001789 0.417668229     No
#> 835   0.68785723 0.386476372    Yes
#> 836   0.74286368 0.507571179     No
#> 837   0.68987626 0.275461530    Yes
#> 838   0.60253223 0.136451178     No
#> 839   0.13787241 0.411520463    Yes
#> 840   0.41399654 0.595210123     No
#> 841   0.66618660 0.261146491    Yes
#> 842   0.16747143 0.472149954     No
#> 843   0.36363378 0.365655581    Yes
#> 844   0.72945052 0.389157147     No
#> 845   0.40211495 0.135884730    Yes
#> 846   0.87177815 0.006516951     No
#> 847   0.49172979 0.042850908    Yes
#> 848   0.66353534 0.012354141     No
#> 849   0.54816316 0.186607659    Yes
#> 850   0.43927018 0.195573613     No
#> 851   0.58766825 0.539660962    Yes
#> 852   0.19372644 0.461853811     No
#> 853   0.36424137 0.071836387    Yes
#> 854   0.60268532 0.217212784     No
#> 855   0.50814950 0.160573053    Yes
#> 856   0.57403088 0.535286790     No
#> 857   0.80891580 0.063551108    Yes
#> 858   0.66080555 0.207301304     No
#> 859   0.35262665 0.131934818    Yes
#> 860   0.54119871 0.327473600     No
#> 861   0.52045849 0.126929427    Yes
#> 862   0.18797580 0.593378821     No
#> 863   0.68585242 0.372396663    Yes
#> 864   0.99616135 0.318422948     No
#> 865   0.55802169 0.195180455    Yes
#> 866   0.61090853 0.232337822     No
#> 867   0.25017721 0.425583519    Yes
#> 868   0.39202214 0.207503032     No
#> 869   0.56014276 0.372323025    Yes
#> 870   0.40964296 0.454286314     No
#> 871   0.43422640 0.132933513    Yes
#> 872   0.66875441 0.011974596     No
#> 873   0.40035979 0.407187511    Yes
#> 874   0.67225415 0.378223932     No
#> 875   0.45543231 0.312028499    Yes
#> 876   0.76741789 0.200362143     No
#> 877   0.43224602 0.240926166    Yes
#> 878   0.08529002 0.579631291     No
#> 879   0.41382512 0.465168390    Yes
#> 880   0.97066447 0.328370631     No
#> 881   0.55102534 0.342261183    Yes
#> 882   0.48650110 0.441905645     No
#> 883   0.66342761 0.555620749    Yes
#> 884   0.73913575 0.309384938     No
#> 885  -0.15102802 0.327702073    Yes
#> 886   0.77917734 0.450656877     No
#> 887   0.61397708 0.249915379    Yes
#> 888   0.82961103 0.370494035     No
#> 889   0.61947372 0.277727210    Yes
#> 890   0.55337617 0.095705962     No
#> 891   0.26634638 0.014462903    Yes
#> 892   0.36279149 0.459377337     No
#> 893   0.77610074 0.499797150    Yes
#> 894   0.97343609 0.036536516     No
#> 895   0.61348425 0.285210663    Yes
#> 896   0.05957072 0.053203518     No
#> 897   0.55377182 0.473441250    Yes
#> 898   0.35417790 0.545332176     No
#> 899   0.54177183 0.592745642    Yes
#> 900   0.17425675 0.176117167     No
#> 901   0.28490898 0.134423125    Yes
#> 902   0.24278264 0.374849713     No
#> 903   0.25794322 0.471348277    Yes
#> 904   0.51756041 0.501435921     No
#> 905   0.40260229 0.471717555    Yes
#> 906   0.97898856 0.041919373     No
#> 907   0.37630519 0.270802170    Yes
#> 908   0.15030532 0.582584340     No
#> 909   0.65094469 0.436970887    Yes
#> 910   0.44144997 0.010409170     No
#> 911   0.70782421 0.096153317    Yes
#> 912   0.29768528 0.355558874     No
#> 913   0.57683552 0.466442367    Yes
#> 914   0.28672801 0.056129793     No
#> 915   0.48073405 0.099503653    Yes
#> 916   0.76800428 0.105509859     No
#> 917   0.52065851 0.451008492    Yes
#> 918   0.29753286 0.542230731     No
#> 919   0.60576005 0.217237120    Yes
#> 920   0.55740152 0.097001268     No
#> 921   0.40078678 0.266465657    Yes
#> 922   0.19936941 0.514106114     No
#> 923   0.42644088 0.112576448    Yes
#> 924   0.41888542 0.552212434     No
#> 925   0.49520211 0.237739326    Yes
#> 926   0.34117591 0.011898441     No
#> 927   0.50977338 0.202322198    Yes
#> 928   0.56275339 0.117395468     No
#> 929   0.66183151 0.115626798    Yes
#> 930   0.31800476 0.013107151     No
#> 931   0.41687145 0.475938666    Yes
#> 932   0.31804226 0.297385642     No
#> 933   0.37867445 0.120684234    Yes
#> 934   0.76001789 0.417668229     No
#> 935   0.68785723 0.386476372    Yes
#> 936   0.74286368 0.507571179     No
#> 937   0.68987626 0.275461530    Yes
#> 938   0.60253223 0.136451178     No
#> 939   0.13787241 0.411520463    Yes
#> 940   0.41399654 0.595210123     No
#> 941   0.66618660 0.261146491    Yes
#> 942   0.16747143 0.472149954     No
#> 943   0.36363378 0.365655581    Yes
#> 944   0.72945052 0.389157147     No
#> 945   0.40211495 0.135884730    Yes
#> 946   0.87177815 0.006516951     No
#> 947   0.49172979 0.042850908    Yes
#> 948   0.66353534 0.012354141     No
#> 949   0.54816316 0.186607659    Yes
#> 950   0.43927018 0.195573613     No
#> 951   0.58766825 0.539660962    Yes
#> 952   0.19372644 0.461853811     No
#> 953   0.36424137 0.071836387    Yes
#> 954   0.60268532 0.217212784     No
#> 955   0.50814950 0.160573053    Yes
#> 956   0.57403088 0.535286790     No
#> 957   0.80891580 0.063551108    Yes
#> 958   0.66080555 0.207301304     No
#> 959   0.35262665 0.131934818    Yes
#> 960   0.54119871 0.327473600     No
#> 961   0.52045849 0.126929427    Yes
#> 962   0.18797580 0.593378821     No
#> 963   0.68585242 0.372396663    Yes
#> 964   0.99616135 0.318422948     No
#> 965   0.55802169 0.195180455    Yes
#> 966   0.61090853 0.232337822     No
#> 967   0.25017721 0.425583519    Yes
#> 968   0.39202214 0.207503032     No
#> 969   0.56014276 0.372323025    Yes
#> 970   0.40964296 0.454286314     No
#> 971   0.43422640 0.132933513    Yes
#> 972   0.66875441 0.011974596     No
#> 973   0.40035979 0.407187511    Yes
#> 974   0.67225415 0.378223932     No
#> 975   0.45543231 0.312028499    Yes
#> 976   0.76741789 0.200362143     No
#> 977   0.43224602 0.240926166    Yes
#> 978   0.08529002 0.579631291     No
#> 979   0.41382512 0.465168390    Yes
#> 980   0.97066447 0.328370631     No
#> 981   0.55102534 0.342261183    Yes
#> 982   0.48650110 0.441905645     No
#> 983   0.66342761 0.555620749    Yes
#> 984   0.73913575 0.309384938     No
#> 985  -0.15102802 0.327702073    Yes
#> 986   0.77917734 0.450656877     No
#> 987   0.61397708 0.249915379    Yes
#> 988   0.82961103 0.370494035     No
#> 989   0.61947372 0.277727210    Yes
#> 990   0.55337617 0.095705962     No
#> 991   0.26634638 0.014462903    Yes
#> 992   0.36279149 0.459377337     No
#> 993   0.77610074 0.499797150    Yes
#> 994   0.97343609 0.036536516     No
#> 995   0.61348425 0.285210663    Yes
#> 996   0.05957072 0.053203518     No
#> 997   0.55377182 0.473441250    Yes
#> 998   0.35417790 0.545332176     No
#> 999   0.54177183 0.592745642    Yes
#> 1000  0.17425675 0.176117167     No

# A pipe friendly example
if (FALSE) {
library(dplyr)
as_tibble(yields) %>%
select_col(height, weight, normal)
}