Get row differences between values

This function returns the differences between rows depending on the user's choice.

rowdiff(
  df,
  direction = "forward",
  exclude = NULL,
  na.rm = FALSE,
  na_action = NULL,
  ...
)

Arguments

df
The data set for which differences are required

direction
One of forward and reverse. The default is forward meaning the differences are calculated in such a way that the difference between the current value and the next is returned

exclude
A character vector specifying what classes should be removed. See examples below

na.rm
Logical. Should missing values be removed? The missing values referred to are those introduced during the calculation ie when subtracting a row with itself. Defaults to FALSE.

na_action
If na.rm is TRUE, how should missing values be replaced? Depending on the value as set out in `na_replace`, the value can be replaced as per the user's requirement.

...
Other arguments to `na_replace`.

Value

A data.frame object of row differences

See also

Examples

# Remove factor columns
data("yields", package="manymodelr")
rowdiff(yields,exclude = "factor",direction = "reverse")

#>            height       weight       yield
#> 1              NA           NA          NA
#> 2    -0.042126342  0.240426589 -15.8083025
#> 3     0.015160589  0.096498564  11.1708254
#> 4     0.259617182  0.030087643   6.5784242
#> 5    -0.114958113 -0.029718366 -19.5840896
#> 6     0.576386269 -0.429798182   6.8257190
#> 7    -0.602683375  0.228882797  19.3698408
#> 8    -0.225999864  0.311782170 -17.0859896
#> 9     0.500639373 -0.145613453   2.9720687
#> 10   -0.209494722 -0.426561716  -6.3184209
#> 11    0.266374238  0.085744146  27.1245118
#> 12   -0.410138928  0.259405558 -15.2716605
#> 13    0.279150234  0.110883493  -0.2339727
#> 14   -0.290107507 -0.410312574   1.3784473
#> 15    0.194006044  0.043373860  13.8365258
#> 16    0.287270225  0.006006207 -25.5056162
#> 17   -0.247345772  0.345498632  23.1938603
#> 18   -0.223125646  0.091222239  -3.7165462
#> 19    0.308227185 -0.324993611 -14.6373280
#> 20   -0.048358525 -0.120235852  -4.4740984
#> 21   -0.156614736  0.169464389  16.3206098
#> 22   -0.201417377  0.247640457  -5.6548240
#> 23    0.227071468 -0.401529667   2.2748754
#> 24   -0.007555453  0.439635987  -0.1595699
#> 25    0.076316690 -0.314473108 -16.7198606
#> 26   -0.154026208 -0.225840885  22.9991098
#> 27    0.168597472  0.190423758  -5.9523314
#> 28    0.052980017 -0.084926730  -2.0215337
#> 29    0.099078114 -0.001768670  -8.9893929
#> 30   -0.343826751 -0.102519647  -5.0617928
#> 31    0.098866692  0.462831515  12.4706444
#> 32   -0.098829186 -0.178553024  15.9069312
#> 33    0.060632189 -0.176701408 -15.6250145
#> 34    0.381343433  0.296983995  -3.1182104
#> 35   -0.072160660 -0.031191856  -2.6539223
#> 36    0.055006456  0.121094807 -14.9181948
#> 37   -0.052987426 -0.232109650  10.3043758
#> 38   -0.087344022 -0.139010352   5.2136938
#> 39   -0.464659827  0.275069285  -5.6574819
#> 40    0.276124135  0.183689660  11.2296970
#> 41    0.252190060 -0.334063632 -14.0395285
#> 42   -0.498715175  0.211003463  -1.4726871
#> 43    0.196162356 -0.106494373  22.5695662
#> 44    0.365816733  0.023501566 -18.1036941
#> 45   -0.327335562 -0.253272418  14.2016053
#> 46    0.469663200 -0.129367779 -11.0213202
#> 47   -0.380048363  0.036333957   2.1232694
#> 48    0.171805546 -0.030496767   6.1191398
#> 49   -0.115372180  0.174253518   5.9530589
#> 50   -0.108892981  0.008965954   5.8287910
#> 51    0.148398072  0.344087349  -8.4141645
#> 52   -0.393941807 -0.077807151  -7.0315408
#> 53    0.170514924 -0.390017424  10.1251790
#> 54    0.238443951  0.145376397   9.7358719
#> 55   -0.094535819 -0.056639732 -20.1823975
#> 56    0.065881376  0.374713737   2.2348351
#> 57    0.234884927 -0.471735681   5.8700792
#> 58   -0.148110250  0.143750196 -16.1936983
#> 59   -0.308178899 -0.075366487  -2.8807672
#> 60    0.188572057  0.195538782  15.6524780
#> 61   -0.020740220 -0.200544173   0.9213956
#> 62   -0.332482690  0.466449393  -3.3260430
#> 63    0.497876617 -0.220982158  10.3977071
#> 64    0.310308933 -0.053973714   3.3022905
#> 65   -0.438139663 -0.123242493  -4.6348181
#> 66    0.052886841  0.037157367 -15.2681432
#> 67   -0.360731318  0.193245697   5.8782861
#> 68    0.141844932 -0.218080487  18.3319071
#> 69    0.168120615  0.164819993 -15.3172639
#> 70   -0.150499798  0.081963290   9.1542837
#> 71    0.024583441 -0.321352802 -15.2441627
#> 72    0.234528014 -0.120958917  -5.2507718
#> 73   -0.268394621  0.395212915  10.8397037
#> 74    0.271894354 -0.028963579 -32.2443997
#> 75   -0.216821832 -0.066195433  20.4521392
#> 76    0.311985574 -0.111666356  20.2597409
#> 77   -0.335171869  0.040564023 -11.2149940
#> 78   -0.346956004  0.338705125   9.8190935
#> 79    0.328535109 -0.114462901 -20.9018006
#> 80    0.556839345 -0.136797758   8.4664328
#> 81   -0.419639128  0.013890552   9.6341730
#> 82   -0.064524237  0.099644461   8.8114199
#> 83    0.176926501  0.113715105 -15.9523425
#> 84    0.075708146 -0.246235811   1.8527718
#> 85   -0.890163772  0.018317135   1.7636756
#> 86    0.930205362  0.122954804  15.3722437
#> 87   -0.165200263 -0.200741498 -11.7759961
#> 88    0.215633949  0.120578657  -2.8702685
#> 89   -0.210137306 -0.092766825  11.5393091
#> 90   -0.066097555 -0.182021248  -1.2980116
#> 91   -0.287029791 -0.081243060 -15.8531082
#> 92    0.096445113  0.444914434   4.7521878
#> 93    0.413309252  0.040419813  15.3440911
#> 94    0.197335351 -0.463260634 -18.3280455
#> 95   -0.359951846  0.248674147   3.3711593
#> 96   -0.553913527 -0.232007145  -0.6571838
#> 97    0.494201101  0.420237732  -9.3034516
#> 98   -0.199593915  0.071890926  10.1017754
#> 99    0.187593927  0.047413466   4.1068136
#> 100  -0.367515081 -0.416628475 -24.7481263
#> 101   0.110652227 -0.041694042  18.6402481
#> 102  -0.042126342  0.240426589 -15.8083025
#> 103   0.015160589  0.096498564  11.1708254
#> 104   0.259617182  0.030087643   6.5784242
#> 105  -0.114958113 -0.029718366 -19.5840896
#> 106   0.576386269 -0.429798182   6.8257190
#> 107  -0.602683375  0.228882797  19.3698408
#> 108  -0.225999864  0.311782170 -17.0859896
#> 109   0.500639373 -0.145613453   2.9720687
#> 110  -0.209494722 -0.426561716  -6.3184209
#> 111   0.266374238  0.085744146  27.1245118
#> 112  -0.410138928  0.259405558 -15.2716605
#> 113   0.279150234  0.110883493  -0.2339727
#> 114  -0.290107507 -0.410312574   1.3784473
#> 115   0.194006044  0.043373860  13.8365258
#> 116   0.287270225  0.006006207 -25.5056162
#> 117  -0.247345772  0.345498632  23.1938603
#> 118  -0.223125646  0.091222239  -3.7165462
#> 119   0.308227185 -0.324993611 -14.6373280
#> 120  -0.048358525 -0.120235852  -4.4740984
#> 121  -0.156614736  0.169464389  16.3206098
#> 122  -0.201417377  0.247640457  -5.6548240
#> 123   0.227071468 -0.401529667   2.2748754
#> 124  -0.007555453  0.439635987  -0.1595699
#> 125   0.076316690 -0.314473108 -16.7198606
#> 126  -0.154026208 -0.225840885  22.9991098
#> 127   0.168597472  0.190423758  -5.9523314
#> 128   0.052980017 -0.084926730  -2.0215337
#> 129   0.099078114 -0.001768670  -8.9893929
#> 130  -0.343826751 -0.102519647  -5.0617928
#> 131   0.098866692  0.462831515  12.4706444
#> 132  -0.098829186 -0.178553024  15.9069312
#> 133   0.060632189 -0.176701408 -15.6250145
#> 134   0.381343433  0.296983995  -3.1182104
#> 135  -0.072160660 -0.031191856  -2.6539223
#> 136   0.055006456  0.121094807 -14.9181948
#> 137  -0.052987426 -0.232109650  10.3043758
#> 138  -0.087344022 -0.139010352   5.2136938
#> 139  -0.464659827  0.275069285  -5.6574819
#> 140   0.276124135  0.183689660  11.2296970
#> 141   0.252190060 -0.334063632 -14.0395285
#> 142  -0.498715175  0.211003463  -1.4726871
#> 143   0.196162356 -0.106494373  22.5695662
#> 144   0.365816733  0.023501566 -18.1036941
#> 145  -0.327335562 -0.253272418  14.2016053
#> 146   0.469663200 -0.129367779 -11.0213202
#> 147  -0.380048363  0.036333957   2.1232694
#> 148   0.171805546 -0.030496767   6.1191398
#> 149  -0.115372180  0.174253518   5.9530589
#> 150  -0.108892981  0.008965954   5.8287910
#> 151   0.148398072  0.344087349  -8.4141645
#> 152  -0.393941807 -0.077807151  -7.0315408
#> 153   0.170514924 -0.390017424  10.1251790
#> 154   0.238443951  0.145376397   9.7358719
#> 155  -0.094535819 -0.056639732 -20.1823975
#> 156   0.065881376  0.374713737   2.2348351
#> 157   0.234884927 -0.471735681   5.8700792
#> 158  -0.148110250  0.143750196 -16.1936983
#> 159  -0.308178899 -0.075366487  -2.8807672
#> 160   0.188572057  0.195538782  15.6524780
#> 161  -0.020740220 -0.200544173   0.9213956
#> 162  -0.332482690  0.466449393  -3.3260430
#> 163   0.497876617 -0.220982158  10.3977071
#> 164   0.310308933 -0.053973714   3.3022905
#> 165  -0.438139663 -0.123242493  -4.6348181
#> 166   0.052886841  0.037157367 -15.2681432
#> 167  -0.360731318  0.193245697   5.8782861
#> 168   0.141844932 -0.218080487  18.3319071
#> 169   0.168120615  0.164819993 -15.3172639
#> 170  -0.150499798  0.081963290   9.1542837
#> 171   0.024583441 -0.321352802 -15.2441627
#> 172   0.234528014 -0.120958917  -5.2507718
#> 173  -0.268394621  0.395212915  10.8397037
#> 174   0.271894354 -0.028963579 -32.2443997
#> 175  -0.216821832 -0.066195433  20.4521392
#> 176   0.311985574 -0.111666356  20.2597409
#> 177  -0.335171869  0.040564023 -11.2149940
#> 178  -0.346956004  0.338705125   9.8190935
#> 179   0.328535109 -0.114462901 -20.9018006
#> 180   0.556839345 -0.136797758   8.4664328
#> 181  -0.419639128  0.013890552   9.6341730
#> 182  -0.064524237  0.099644461   8.8114199
#> 183   0.176926501  0.113715105 -15.9523425
#> 184   0.075708146 -0.246235811   1.8527718
#> 185  -0.890163772  0.018317135   1.7636756
#> 186   0.930205362  0.122954804  15.3722437
#> 187  -0.165200263 -0.200741498 -11.7759961
#> 188   0.215633949  0.120578657  -2.8702685
#> 189  -0.210137306 -0.092766825  11.5393091
#> 190  -0.066097555 -0.182021248  -1.2980116
#> 191  -0.287029791 -0.081243060 -15.8531082
#> 192   0.096445113  0.444914434   4.7521878
#> 193   0.413309252  0.040419813  15.3440911
#> 194   0.197335351 -0.463260634 -18.3280455
#> 195  -0.359951846  0.248674147   3.3711593
#> 196  -0.553913527 -0.232007145  -0.6571838
#> 197   0.494201101  0.420237732  -9.3034516
#> 198  -0.199593915  0.071890926  10.1017754
#> 199   0.187593927  0.047413466   4.1068136
#> 200  -0.367515081 -0.416628475 -24.7481263
#> 201   0.110652227 -0.041694042  18.6402481
#> 202  -0.042126342  0.240426589 -15.8083025
#> 203   0.015160589  0.096498564  11.1708254
#> 204   0.259617182  0.030087643   6.5784242
#> 205  -0.114958113 -0.029718366 -19.5840896
#> 206   0.576386269 -0.429798182   6.8257190
#> 207  -0.602683375  0.228882797  19.3698408
#> 208  -0.225999864  0.311782170 -17.0859896
#> 209   0.500639373 -0.145613453   2.9720687
#> 210  -0.209494722 -0.426561716  -6.3184209
#> 211   0.266374238  0.085744146  27.1245118
#> 212  -0.410138928  0.259405558 -15.2716605
#> 213   0.279150234  0.110883493  -0.2339727
#> 214  -0.290107507 -0.410312574   1.3784473
#> 215   0.194006044  0.043373860  13.8365258
#> 216   0.287270225  0.006006207 -25.5056162
#> 217  -0.247345772  0.345498632  23.1938603
#> 218  -0.223125646  0.091222239  -3.7165462
#> 219   0.308227185 -0.324993611 -14.6373280
#> 220  -0.048358525 -0.120235852  -4.4740984
#> 221  -0.156614736  0.169464389  16.3206098
#> 222  -0.201417377  0.247640457  -5.6548240
#> 223   0.227071468 -0.401529667   2.2748754
#> 224  -0.007555453  0.439635987  -0.1595699
#> 225   0.076316690 -0.314473108 -16.7198606
#> 226  -0.154026208 -0.225840885  22.9991098
#> 227   0.168597472  0.190423758  -5.9523314
#> 228   0.052980017 -0.084926730  -2.0215337
#> 229   0.099078114 -0.001768670  -8.9893929
#> 230  -0.343826751 -0.102519647  -5.0617928
#> 231   0.098866692  0.462831515  12.4706444
#> 232  -0.098829186 -0.178553024  15.9069312
#> 233   0.060632189 -0.176701408 -15.6250145
#> 234   0.381343433  0.296983995  -3.1182104
#> 235  -0.072160660 -0.031191856  -2.6539223
#> 236   0.055006456  0.121094807 -14.9181948
#> 237  -0.052987426 -0.232109650  10.3043758
#> 238  -0.087344022 -0.139010352   5.2136938
#> 239  -0.464659827  0.275069285  -5.6574819
#> 240   0.276124135  0.183689660  11.2296970
#> 241   0.252190060 -0.334063632 -14.0395285
#> 242  -0.498715175  0.211003463  -1.4726871
#> 243   0.196162356 -0.106494373  22.5695662
#> 244   0.365816733  0.023501566 -18.1036941
#> 245  -0.327335562 -0.253272418  14.2016053
#> 246   0.469663200 -0.129367779 -11.0213202
#> 247  -0.380048363  0.036333957   2.1232694
#> 248   0.171805546 -0.030496767   6.1191398
#> 249  -0.115372180  0.174253518   5.9530589
#> 250  -0.108892981  0.008965954   5.8287910
#> 251   0.148398072  0.344087349  -8.4141645
#> 252  -0.393941807 -0.077807151  -7.0315408
#> 253   0.170514924 -0.390017424  10.1251790
#> 254   0.238443951  0.145376397   9.7358719
#> 255  -0.094535819 -0.056639732 -20.1823975
#> 256   0.065881376  0.374713737   2.2348351
#> 257   0.234884927 -0.471735681   5.8700792
#> 258  -0.148110250  0.143750196 -16.1936983
#> 259  -0.308178899 -0.075366487  -2.8807672
#> 260   0.188572057  0.195538782  15.6524780
#> 261  -0.020740220 -0.200544173   0.9213956
#> 262  -0.332482690  0.466449393  -3.3260430
#> 263   0.497876617 -0.220982158  10.3977071
#> 264   0.310308933 -0.053973714   3.3022905
#> 265  -0.438139663 -0.123242493  -4.6348181
#> 266   0.052886841  0.037157367 -15.2681432
#> 267  -0.360731318  0.193245697   5.8782861
#> 268   0.141844932 -0.218080487  18.3319071
#> 269   0.168120615  0.164819993 -15.3172639
#> 270  -0.150499798  0.081963290   9.1542837
#> 271   0.024583441 -0.321352802 -15.2441627
#> 272   0.234528014 -0.120958917  -5.2507718
#> 273  -0.268394621  0.395212915  10.8397037
#> 274   0.271894354 -0.028963579 -32.2443997
#> 275  -0.216821832 -0.066195433  20.4521392
#> 276   0.311985574 -0.111666356  20.2597409
#> 277  -0.335171869  0.040564023 -11.2149940
#> 278  -0.346956004  0.338705125   9.8190935
#> 279   0.328535109 -0.114462901 -20.9018006
#> 280   0.556839345 -0.136797758   8.4664328
#> 281  -0.419639128  0.013890552   9.6341730
#> 282  -0.064524237  0.099644461   8.8114199
#> 283   0.176926501  0.113715105 -15.9523425
#> 284   0.075708146 -0.246235811   1.8527718
#> 285  -0.890163772  0.018317135   1.7636756
#> 286   0.930205362  0.122954804  15.3722437
#> 287  -0.165200263 -0.200741498 -11.7759961
#> 288   0.215633949  0.120578657  -2.8702685
#> 289  -0.210137306 -0.092766825  11.5393091
#> 290  -0.066097555 -0.182021248  -1.2980116
#> 291  -0.287029791 -0.081243060 -15.8531082
#> 292   0.096445113  0.444914434   4.7521878
#> 293   0.413309252  0.040419813  15.3440911
#> 294   0.197335351 -0.463260634 -18.3280455
#> 295  -0.359951846  0.248674147   3.3711593
#> 296  -0.553913527 -0.232007145  -0.6571838
#> 297   0.494201101  0.420237732  -9.3034516
#> 298  -0.199593915  0.071890926  10.1017754
#> 299   0.187593927  0.047413466   4.1068136
#> 300  -0.367515081 -0.416628475 -24.7481263
#> 301   0.110652227 -0.041694042  18.6402481
#> 302  -0.042126342  0.240426589 -15.8083025
#> 303   0.015160589  0.096498564  11.1708254
#> 304   0.259617182  0.030087643   6.5784242
#> 305  -0.114958113 -0.029718366 -19.5840896
#> 306   0.576386269 -0.429798182   6.8257190
#> 307  -0.602683375  0.228882797  19.3698408
#> 308  -0.225999864  0.311782170 -17.0859896
#> 309   0.500639373 -0.145613453   2.9720687
#> 310  -0.209494722 -0.426561716  -6.3184209
#> 311   0.266374238  0.085744146  27.1245118
#> 312  -0.410138928  0.259405558 -15.2716605
#> 313   0.279150234  0.110883493  -0.2339727
#> 314  -0.290107507 -0.410312574   1.3784473
#> 315   0.194006044  0.043373860  13.8365258
#> 316   0.287270225  0.006006207 -25.5056162
#> 317  -0.247345772  0.345498632  23.1938603
#> 318  -0.223125646  0.091222239  -3.7165462
#> 319   0.308227185 -0.324993611 -14.6373280
#> 320  -0.048358525 -0.120235852  -4.4740984
#> 321  -0.156614736  0.169464389  16.3206098
#> 322  -0.201417377  0.247640457  -5.6548240
#> 323   0.227071468 -0.401529667   2.2748754
#> 324  -0.007555453  0.439635987  -0.1595699
#> 325   0.076316690 -0.314473108 -16.7198606
#> 326  -0.154026208 -0.225840885  22.9991098
#> 327   0.168597472  0.190423758  -5.9523314
#> 328   0.052980017 -0.084926730  -2.0215337
#> 329   0.099078114 -0.001768670  -8.9893929
#> 330  -0.343826751 -0.102519647  -5.0617928
#> 331   0.098866692  0.462831515  12.4706444
#> 332  -0.098829186 -0.178553024  15.9069312
#> 333   0.060632189 -0.176701408 -15.6250145
#> 334   0.381343433  0.296983995  -3.1182104
#> 335  -0.072160660 -0.031191856  -2.6539223
#> 336   0.055006456  0.121094807 -14.9181948
#> 337  -0.052987426 -0.232109650  10.3043758
#> 338  -0.087344022 -0.139010352   5.2136938
#> 339  -0.464659827  0.275069285  -5.6574819
#> 340   0.276124135  0.183689660  11.2296970
#> 341   0.252190060 -0.334063632 -14.0395285
#> 342  -0.498715175  0.211003463  -1.4726871
#> 343   0.196162356 -0.106494373  22.5695662
#> 344   0.365816733  0.023501566 -18.1036941
#> 345  -0.327335562 -0.253272418  14.2016053
#> 346   0.469663200 -0.129367779 -11.0213202
#> 347  -0.380048363  0.036333957   2.1232694
#> 348   0.171805546 -0.030496767   6.1191398
#> 349  -0.115372180  0.174253518   5.9530589
#> 350  -0.108892981  0.008965954   5.8287910
#> 351   0.148398072  0.344087349  -8.4141645
#> 352  -0.393941807 -0.077807151  -7.0315408
#> 353   0.170514924 -0.390017424  10.1251790
#> 354   0.238443951  0.145376397   9.7358719
#> 355  -0.094535819 -0.056639732 -20.1823975
#> 356   0.065881376  0.374713737   2.2348351
#> 357   0.234884927 -0.471735681   5.8700792
#> 358  -0.148110250  0.143750196 -16.1936983
#> 359  -0.308178899 -0.075366487  -2.8807672
#> 360   0.188572057  0.195538782  15.6524780
#> 361  -0.020740220 -0.200544173   0.9213956
#> 362  -0.332482690  0.466449393  -3.3260430
#> 363   0.497876617 -0.220982158  10.3977071
#> 364   0.310308933 -0.053973714   3.3022905
#> 365  -0.438139663 -0.123242493  -4.6348181
#> 366   0.052886841  0.037157367 -15.2681432
#> 367  -0.360731318  0.193245697   5.8782861
#> 368   0.141844932 -0.218080487  18.3319071
#> 369   0.168120615  0.164819993 -15.3172639
#> 370  -0.150499798  0.081963290   9.1542837
#> 371   0.024583441 -0.321352802 -15.2441627
#> 372   0.234528014 -0.120958917  -5.2507718
#> 373  -0.268394621  0.395212915  10.8397037
#> 374   0.271894354 -0.028963579 -32.2443997
#> 375  -0.216821832 -0.066195433  20.4521392
#> 376   0.311985574 -0.111666356  20.2597409
#> 377  -0.335171869  0.040564023 -11.2149940
#> 378  -0.346956004  0.338705125   9.8190935
#> 379   0.328535109 -0.114462901 -20.9018006
#> 380   0.556839345 -0.136797758   8.4664328
#> 381  -0.419639128  0.013890552   9.6341730
#> 382  -0.064524237  0.099644461   8.8114199
#> 383   0.176926501  0.113715105 -15.9523425
#> 384   0.075708146 -0.246235811   1.8527718
#> 385  -0.890163772  0.018317135   1.7636756
#> 386   0.930205362  0.122954804  15.3722437
#> 387  -0.165200263 -0.200741498 -11.7759961
#> 388   0.215633949  0.120578657  -2.8702685
#> 389  -0.210137306 -0.092766825  11.5393091
#> 390  -0.066097555 -0.182021248  -1.2980116
#> 391  -0.287029791 -0.081243060 -15.8531082
#> 392   0.096445113  0.444914434   4.7521878
#> 393   0.413309252  0.040419813  15.3440911
#> 394   0.197335351 -0.463260634 -18.3280455
#> 395  -0.359951846  0.248674147   3.3711593
#> 396  -0.553913527 -0.232007145  -0.6571838
#> 397   0.494201101  0.420237732  -9.3034516
#> 398  -0.199593915  0.071890926  10.1017754
#> 399   0.187593927  0.047413466   4.1068136
#> 400  -0.367515081 -0.416628475 -24.7481263
#> 401   0.110652227 -0.041694042  18.6402481
#> 402  -0.042126342  0.240426589 -15.8083025
#> 403   0.015160589  0.096498564  11.1708254
#> 404   0.259617182  0.030087643   6.5784242
#> 405  -0.114958113 -0.029718366 -19.5840896
#> 406   0.576386269 -0.429798182   6.8257190
#> 407  -0.602683375  0.228882797  19.3698408
#> 408  -0.225999864  0.311782170 -17.0859896
#> 409   0.500639373 -0.145613453   2.9720687
#> 410  -0.209494722 -0.426561716  -6.3184209
#> 411   0.266374238  0.085744146  27.1245118
#> 412  -0.410138928  0.259405558 -15.2716605
#> 413   0.279150234  0.110883493  -0.2339727
#> 414  -0.290107507 -0.410312574   1.3784473
#> 415   0.194006044  0.043373860  13.8365258
#> 416   0.287270225  0.006006207 -25.5056162
#> 417  -0.247345772  0.345498632  23.1938603
#> 418  -0.223125646  0.091222239  -3.7165462
#> 419   0.308227185 -0.324993611 -14.6373280
#> 420  -0.048358525 -0.120235852  -4.4740984
#> 421  -0.156614736  0.169464389  16.3206098
#> 422  -0.201417377  0.247640457  -5.6548240
#> 423   0.227071468 -0.401529667   2.2748754
#> 424  -0.007555453  0.439635987  -0.1595699
#> 425   0.076316690 -0.314473108 -16.7198606
#> 426  -0.154026208 -0.225840885  22.9991098
#> 427   0.168597472  0.190423758  -5.9523314
#> 428   0.052980017 -0.084926730  -2.0215337
#> 429   0.099078114 -0.001768670  -8.9893929
#> 430  -0.343826751 -0.102519647  -5.0617928
#> 431   0.098866692  0.462831515  12.4706444
#> 432  -0.098829186 -0.178553024  15.9069312
#> 433   0.060632189 -0.176701408 -15.6250145
#> 434   0.381343433  0.296983995  -3.1182104
#> 435  -0.072160660 -0.031191856  -2.6539223
#> 436   0.055006456  0.121094807 -14.9181948
#> 437  -0.052987426 -0.232109650  10.3043758
#> 438  -0.087344022 -0.139010352   5.2136938
#> 439  -0.464659827  0.275069285  -5.6574819
#> 440   0.276124135  0.183689660  11.2296970
#> 441   0.252190060 -0.334063632 -14.0395285
#> 442  -0.498715175  0.211003463  -1.4726871
#> 443   0.196162356 -0.106494373  22.5695662
#> 444   0.365816733  0.023501566 -18.1036941
#> 445  -0.327335562 -0.253272418  14.2016053
#> 446   0.469663200 -0.129367779 -11.0213202
#> 447  -0.380048363  0.036333957   2.1232694
#> 448   0.171805546 -0.030496767   6.1191398
#> 449  -0.115372180  0.174253518   5.9530589
#> 450  -0.108892981  0.008965954   5.8287910
#> 451   0.148398072  0.344087349  -8.4141645
#> 452  -0.393941807 -0.077807151  -7.0315408
#> 453   0.170514924 -0.390017424  10.1251790
#> 454   0.238443951  0.145376397   9.7358719
#> 455  -0.094535819 -0.056639732 -20.1823975
#> 456   0.065881376  0.374713737   2.2348351
#> 457   0.234884927 -0.471735681   5.8700792
#> 458  -0.148110250  0.143750196 -16.1936983
#> 459  -0.308178899 -0.075366487  -2.8807672
#> 460   0.188572057  0.195538782  15.6524780
#> 461  -0.020740220 -0.200544173   0.9213956
#> 462  -0.332482690  0.466449393  -3.3260430
#> 463   0.497876617 -0.220982158  10.3977071
#> 464   0.310308933 -0.053973714   3.3022905
#> 465  -0.438139663 -0.123242493  -4.6348181
#> 466   0.052886841  0.037157367 -15.2681432
#> 467  -0.360731318  0.193245697   5.8782861
#> 468   0.141844932 -0.218080487  18.3319071
#> 469   0.168120615  0.164819993 -15.3172639
#> 470  -0.150499798  0.081963290   9.1542837
#> 471   0.024583441 -0.321352802 -15.2441627
#> 472   0.234528014 -0.120958917  -5.2507718
#> 473  -0.268394621  0.395212915  10.8397037
#> 474   0.271894354 -0.028963579 -32.2443997
#> 475  -0.216821832 -0.066195433  20.4521392
#> 476   0.311985574 -0.111666356  20.2597409
#> 477  -0.335171869  0.040564023 -11.2149940
#> 478  -0.346956004  0.338705125   9.8190935
#> 479   0.328535109 -0.114462901 -20.9018006
#> 480   0.556839345 -0.136797758   8.4664328
#> 481  -0.419639128  0.013890552   9.6341730
#> 482  -0.064524237  0.099644461   8.8114199
#> 483   0.176926501  0.113715105 -15.9523425
#> 484   0.075708146 -0.246235811   1.8527718
#> 485  -0.890163772  0.018317135   1.7636756
#> 486   0.930205362  0.122954804  15.3722437
#> 487  -0.165200263 -0.200741498 -11.7759961
#> 488   0.215633949  0.120578657  -2.8702685
#> 489  -0.210137306 -0.092766825  11.5393091
#> 490  -0.066097555 -0.182021248  -1.2980116
#> 491  -0.287029791 -0.081243060 -15.8531082
#> 492   0.096445113  0.444914434   4.7521878
#> 493   0.413309252  0.040419813  15.3440911
#> 494   0.197335351 -0.463260634 -18.3280455
#> 495  -0.359951846  0.248674147   3.3711593
#> 496  -0.553913527 -0.232007145  -0.6571838
#> 497   0.494201101  0.420237732  -9.3034516
#> 498  -0.199593915  0.071890926  10.1017754
#> 499   0.187593927  0.047413466   4.1068136
#> 500  -0.367515081 -0.416628475 -24.7481263
#> 501   0.110652227 -0.041694042  18.6402481
#> 502  -0.042126342  0.240426589 -15.8083025
#> 503   0.015160589  0.096498564  11.1708254
#> 504   0.259617182  0.030087643   6.5784242
#> 505  -0.114958113 -0.029718366 -19.5840896
#> 506   0.576386269 -0.429798182   6.8257190
#> 507  -0.602683375  0.228882797  19.3698408
#> 508  -0.225999864  0.311782170 -17.0859896
#> 509   0.500639373 -0.145613453   2.9720687
#> 510  -0.209494722 -0.426561716  -6.3184209
#> 511   0.266374238  0.085744146  27.1245118
#> 512  -0.410138928  0.259405558 -15.2716605
#> 513   0.279150234  0.110883493  -0.2339727
#> 514  -0.290107507 -0.410312574   1.3784473
#> 515   0.194006044  0.043373860  13.8365258
#> 516   0.287270225  0.006006207 -25.5056162
#> 517  -0.247345772  0.345498632  23.1938603
#> 518  -0.223125646  0.091222239  -3.7165462
#> 519   0.308227185 -0.324993611 -14.6373280
#> 520  -0.048358525 -0.120235852  -4.4740984
#> 521  -0.156614736  0.169464389  16.3206098
#> 522  -0.201417377  0.247640457  -5.6548240
#> 523   0.227071468 -0.401529667   2.2748754
#> 524  -0.007555453  0.439635987  -0.1595699
#> 525   0.076316690 -0.314473108 -16.7198606
#> 526  -0.154026208 -0.225840885  22.9991098
#> 527   0.168597472  0.190423758  -5.9523314
#> 528   0.052980017 -0.084926730  -2.0215337
#> 529   0.099078114 -0.001768670  -8.9893929
#> 530  -0.343826751 -0.102519647  -5.0617928
#> 531   0.098866692  0.462831515  12.4706444
#> 532  -0.098829186 -0.178553024  15.9069312
#> 533   0.060632189 -0.176701408 -15.6250145
#> 534   0.381343433  0.296983995  -3.1182104
#> 535  -0.072160660 -0.031191856  -2.6539223
#> 536   0.055006456  0.121094807 -14.9181948
#> 537  -0.052987426 -0.232109650  10.3043758
#> 538  -0.087344022 -0.139010352   5.2136938
#> 539  -0.464659827  0.275069285  -5.6574819
#> 540   0.276124135  0.183689660  11.2296970
#> 541   0.252190060 -0.334063632 -14.0395285
#> 542  -0.498715175  0.211003463  -1.4726871
#> 543   0.196162356 -0.106494373  22.5695662
#> 544   0.365816733  0.023501566 -18.1036941
#> 545  -0.327335562 -0.253272418  14.2016053
#> 546   0.469663200 -0.129367779 -11.0213202
#> 547  -0.380048363  0.036333957   2.1232694
#> 548   0.171805546 -0.030496767   6.1191398
#> 549  -0.115372180  0.174253518   5.9530589
#> 550  -0.108892981  0.008965954   5.8287910
#> 551   0.148398072  0.344087349  -8.4141645
#> 552  -0.393941807 -0.077807151  -7.0315408
#> 553   0.170514924 -0.390017424  10.1251790
#> 554   0.238443951  0.145376397   9.7358719
#> 555  -0.094535819 -0.056639732 -20.1823975
#> 556   0.065881376  0.374713737   2.2348351
#> 557   0.234884927 -0.471735681   5.8700792
#> 558  -0.148110250  0.143750196 -16.1936983
#> 559  -0.308178899 -0.075366487  -2.8807672
#> 560   0.188572057  0.195538782  15.6524780
#> 561  -0.020740220 -0.200544173   0.9213956
#> 562  -0.332482690  0.466449393  -3.3260430
#> 563   0.497876617 -0.220982158  10.3977071
#> 564   0.310308933 -0.053973714   3.3022905
#> 565  -0.438139663 -0.123242493  -4.6348181
#> 566   0.052886841  0.037157367 -15.2681432
#> 567  -0.360731318  0.193245697   5.8782861
#> 568   0.141844932 -0.218080487  18.3319071
#> 569   0.168120615  0.164819993 -15.3172639
#> 570  -0.150499798  0.081963290   9.1542837
#> 571   0.024583441 -0.321352802 -15.2441627
#> 572   0.234528014 -0.120958917  -5.2507718
#> 573  -0.268394621  0.395212915  10.8397037
#> 574   0.271894354 -0.028963579 -32.2443997
#> 575  -0.216821832 -0.066195433  20.4521392
#> 576   0.311985574 -0.111666356  20.2597409
#> 577  -0.335171869  0.040564023 -11.2149940
#> 578  -0.346956004  0.338705125   9.8190935
#> 579   0.328535109 -0.114462901 -20.9018006
#> 580   0.556839345 -0.136797758   8.4664328
#> 581  -0.419639128  0.013890552   9.6341730
#> 582  -0.064524237  0.099644461   8.8114199
#> 583   0.176926501  0.113715105 -15.9523425
#> 584   0.075708146 -0.246235811   1.8527718
#> 585  -0.890163772  0.018317135   1.7636756
#> 586   0.930205362  0.122954804  15.3722437
#> 587  -0.165200263 -0.200741498 -11.7759961
#> 588   0.215633949  0.120578657  -2.8702685
#> 589  -0.210137306 -0.092766825  11.5393091
#> 590  -0.066097555 -0.182021248  -1.2980116
#> 591  -0.287029791 -0.081243060 -15.8531082
#> 592   0.096445113  0.444914434   4.7521878
#> 593   0.413309252  0.040419813  15.3440911
#> 594   0.197335351 -0.463260634 -18.3280455
#> 595  -0.359951846  0.248674147   3.3711593
#> 596  -0.553913527 -0.232007145  -0.6571838
#> 597   0.494201101  0.420237732  -9.3034516
#> 598  -0.199593915  0.071890926  10.1017754
#> 599   0.187593927  0.047413466   4.1068136
#> 600  -0.367515081 -0.416628475 -24.7481263
#> 601   0.110652227 -0.041694042  18.6402481
#> 602  -0.042126342  0.240426589 -15.8083025
#> 603   0.015160589  0.096498564  11.1708254
#> 604   0.259617182  0.030087643   6.5784242
#> 605  -0.114958113 -0.029718366 -19.5840896
#> 606   0.576386269 -0.429798182   6.8257190
#> 607  -0.602683375  0.228882797  19.3698408
#> 608  -0.225999864  0.311782170 -17.0859896
#> 609   0.500639373 -0.145613453   2.9720687
#> 610  -0.209494722 -0.426561716  -6.3184209
#> 611   0.266374238  0.085744146  27.1245118
#> 612  -0.410138928  0.259405558 -15.2716605
#> 613   0.279150234  0.110883493  -0.2339727
#> 614  -0.290107507 -0.410312574   1.3784473
#> 615   0.194006044  0.043373860  13.8365258
#> 616   0.287270225  0.006006207 -25.5056162
#> 617  -0.247345772  0.345498632  23.1938603
#> 618  -0.223125646  0.091222239  -3.7165462
#> 619   0.308227185 -0.324993611 -14.6373280
#> 620  -0.048358525 -0.120235852  -4.4740984
#> 621  -0.156614736  0.169464389  16.3206098
#> 622  -0.201417377  0.247640457  -5.6548240
#> 623   0.227071468 -0.401529667   2.2748754
#> 624  -0.007555453  0.439635987  -0.1595699
#> 625   0.076316690 -0.314473108 -16.7198606
#> 626  -0.154026208 -0.225840885  22.9991098
#> 627   0.168597472  0.190423758  -5.9523314
#> 628   0.052980017 -0.084926730  -2.0215337
#> 629   0.099078114 -0.001768670  -8.9893929
#> 630  -0.343826751 -0.102519647  -5.0617928
#> 631   0.098866692  0.462831515  12.4706444
#> 632  -0.098829186 -0.178553024  15.9069312
#> 633   0.060632189 -0.176701408 -15.6250145
#> 634   0.381343433  0.296983995  -3.1182104
#> 635  -0.072160660 -0.031191856  -2.6539223
#> 636   0.055006456  0.121094807 -14.9181948
#> 637  -0.052987426 -0.232109650  10.3043758
#> 638  -0.087344022 -0.139010352   5.2136938
#> 639  -0.464659827  0.275069285  -5.6574819
#> 640   0.276124135  0.183689660  11.2296970
#> 641   0.252190060 -0.334063632 -14.0395285
#> 642  -0.498715175  0.211003463  -1.4726871
#> 643   0.196162356 -0.106494373  22.5695662
#> 644   0.365816733  0.023501566 -18.1036941
#> 645  -0.327335562 -0.253272418  14.2016053
#> 646   0.469663200 -0.129367779 -11.0213202
#> 647  -0.380048363  0.036333957   2.1232694
#> 648   0.171805546 -0.030496767   6.1191398
#> 649  -0.115372180  0.174253518   5.9530589
#> 650  -0.108892981  0.008965954   5.8287910
#> 651   0.148398072  0.344087349  -8.4141645
#> 652  -0.393941807 -0.077807151  -7.0315408
#> 653   0.170514924 -0.390017424  10.1251790
#> 654   0.238443951  0.145376397   9.7358719
#> 655  -0.094535819 -0.056639732 -20.1823975
#> 656   0.065881376  0.374713737   2.2348351
#> 657   0.234884927 -0.471735681   5.8700792
#> 658  -0.148110250  0.143750196 -16.1936983
#> 659  -0.308178899 -0.075366487  -2.8807672
#> 660   0.188572057  0.195538782  15.6524780
#> 661  -0.020740220 -0.200544173   0.9213956
#> 662  -0.332482690  0.466449393  -3.3260430
#> 663   0.497876617 -0.220982158  10.3977071
#> 664   0.310308933 -0.053973714   3.3022905
#> 665  -0.438139663 -0.123242493  -4.6348181
#> 666   0.052886841  0.037157367 -15.2681432
#> 667  -0.360731318  0.193245697   5.8782861
#> 668   0.141844932 -0.218080487  18.3319071
#> 669   0.168120615  0.164819993 -15.3172639
#> 670  -0.150499798  0.081963290   9.1542837
#> 671   0.024583441 -0.321352802 -15.2441627
#> 672   0.234528014 -0.120958917  -5.2507718
#> 673  -0.268394621  0.395212915  10.8397037
#> 674   0.271894354 -0.028963579 -32.2443997
#> 675  -0.216821832 -0.066195433  20.4521392
#> 676   0.311985574 -0.111666356  20.2597409
#> 677  -0.335171869  0.040564023 -11.2149940
#> 678  -0.346956004  0.338705125   9.8190935
#> 679   0.328535109 -0.114462901 -20.9018006
#> 680   0.556839345 -0.136797758   8.4664328
#> 681  -0.419639128  0.013890552   9.6341730
#> 682  -0.064524237  0.099644461   8.8114199
#> 683   0.176926501  0.113715105 -15.9523425
#> 684   0.075708146 -0.246235811   1.8527718
#> 685  -0.890163772  0.018317135   1.7636756
#> 686   0.930205362  0.122954804  15.3722437
#> 687  -0.165200263 -0.200741498 -11.7759961
#> 688   0.215633949  0.120578657  -2.8702685
#> 689  -0.210137306 -0.092766825  11.5393091
#> 690  -0.066097555 -0.182021248  -1.2980116
#> 691  -0.287029791 -0.081243060 -15.8531082
#> 692   0.096445113  0.444914434   4.7521878
#> 693   0.413309252  0.040419813  15.3440911
#> 694   0.197335351 -0.463260634 -18.3280455
#> 695  -0.359951846  0.248674147   3.3711593
#> 696  -0.553913527 -0.232007145  -0.6571838
#> 697   0.494201101  0.420237732  -9.3034516
#> 698  -0.199593915  0.071890926  10.1017754
#> 699   0.187593927  0.047413466   4.1068136
#> 700  -0.367515081 -0.416628475 -24.7481263
#> 701   0.110652227 -0.041694042  18.6402481
#> 702  -0.042126342  0.240426589 -15.8083025
#> 703   0.015160589  0.096498564  11.1708254
#> 704   0.259617182  0.030087643   6.5784242
#> 705  -0.114958113 -0.029718366 -19.5840896
#> 706   0.576386269 -0.429798182   6.8257190
#> 707  -0.602683375  0.228882797  19.3698408
#> 708  -0.225999864  0.311782170 -17.0859896
#> 709   0.500639373 -0.145613453   2.9720687
#> 710  -0.209494722 -0.426561716  -6.3184209
#> 711   0.266374238  0.085744146  27.1245118
#> 712  -0.410138928  0.259405558 -15.2716605
#> 713   0.279150234  0.110883493  -0.2339727
#> 714  -0.290107507 -0.410312574   1.3784473
#> 715   0.194006044  0.043373860  13.8365258
#> 716   0.287270225  0.006006207 -25.5056162
#> 717  -0.247345772  0.345498632  23.1938603
#> 718  -0.223125646  0.091222239  -3.7165462
#> 719   0.308227185 -0.324993611 -14.6373280
#> 720  -0.048358525 -0.120235852  -4.4740984
#> 721  -0.156614736  0.169464389  16.3206098
#> 722  -0.201417377  0.247640457  -5.6548240
#> 723   0.227071468 -0.401529667   2.2748754
#> 724  -0.007555453  0.439635987  -0.1595699
#> 725   0.076316690 -0.314473108 -16.7198606
#> 726  -0.154026208 -0.225840885  22.9991098
#> 727   0.168597472  0.190423758  -5.9523314
#> 728   0.052980017 -0.084926730  -2.0215337
#> 729   0.099078114 -0.001768670  -8.9893929
#> 730  -0.343826751 -0.102519647  -5.0617928
#> 731   0.098866692  0.462831515  12.4706444
#> 732  -0.098829186 -0.178553024  15.9069312
#> 733   0.060632189 -0.176701408 -15.6250145
#> 734   0.381343433  0.296983995  -3.1182104
#> 735  -0.072160660 -0.031191856  -2.6539223
#> 736   0.055006456  0.121094807 -14.9181948
#> 737  -0.052987426 -0.232109650  10.3043758
#> 738  -0.087344022 -0.139010352   5.2136938
#> 739  -0.464659827  0.275069285  -5.6574819
#> 740   0.276124135  0.183689660  11.2296970
#> 741   0.252190060 -0.334063632 -14.0395285
#> 742  -0.498715175  0.211003463  -1.4726871
#> 743   0.196162356 -0.106494373  22.5695662
#> 744   0.365816733  0.023501566 -18.1036941
#> 745  -0.327335562 -0.253272418  14.2016053
#> 746   0.469663200 -0.129367779 -11.0213202
#> 747  -0.380048363  0.036333957   2.1232694
#> 748   0.171805546 -0.030496767   6.1191398
#> 749  -0.115372180  0.174253518   5.9530589
#> 750  -0.108892981  0.008965954   5.8287910
#> 751   0.148398072  0.344087349  -8.4141645
#> 752  -0.393941807 -0.077807151  -7.0315408
#> 753   0.170514924 -0.390017424  10.1251790
#> 754   0.238443951  0.145376397   9.7358719
#> 755  -0.094535819 -0.056639732 -20.1823975
#> 756   0.065881376  0.374713737   2.2348351
#> 757   0.234884927 -0.471735681   5.8700792
#> 758  -0.148110250  0.143750196 -16.1936983
#> 759  -0.308178899 -0.075366487  -2.8807672
#> 760   0.188572057  0.195538782  15.6524780
#> 761  -0.020740220 -0.200544173   0.9213956
#> 762  -0.332482690  0.466449393  -3.3260430
#> 763   0.497876617 -0.220982158  10.3977071
#> 764   0.310308933 -0.053973714   3.3022905
#> 765  -0.438139663 -0.123242493  -4.6348181
#> 766   0.052886841  0.037157367 -15.2681432
#> 767  -0.360731318  0.193245697   5.8782861
#> 768   0.141844932 -0.218080487  18.3319071
#> 769   0.168120615  0.164819993 -15.3172639
#> 770  -0.150499798  0.081963290   9.1542837
#> 771   0.024583441 -0.321352802 -15.2441627
#> 772   0.234528014 -0.120958917  -5.2507718
#> 773  -0.268394621  0.395212915  10.8397037
#> 774   0.271894354 -0.028963579 -32.2443997
#> 775  -0.216821832 -0.066195433  20.4521392
#> 776   0.311985574 -0.111666356  20.2597409
#> 777  -0.335171869  0.040564023 -11.2149940
#> 778  -0.346956004  0.338705125   9.8190935
#> 779   0.328535109 -0.114462901 -20.9018006
#> 780   0.556839345 -0.136797758   8.4664328
#> 781  -0.419639128  0.013890552   9.6341730
#> 782  -0.064524237  0.099644461   8.8114199
#> 783   0.176926501  0.113715105 -15.9523425
#> 784   0.075708146 -0.246235811   1.8527718
#> 785  -0.890163772  0.018317135   1.7636756
#> 786   0.930205362  0.122954804  15.3722437
#> 787  -0.165200263 -0.200741498 -11.7759961
#> 788   0.215633949  0.120578657  -2.8702685
#> 789  -0.210137306 -0.092766825  11.5393091
#> 790  -0.066097555 -0.182021248  -1.2980116
#> 791  -0.287029791 -0.081243060 -15.8531082
#> 792   0.096445113  0.444914434   4.7521878
#> 793   0.413309252  0.040419813  15.3440911
#> 794   0.197335351 -0.463260634 -18.3280455
#> 795  -0.359951846  0.248674147   3.3711593
#> 796  -0.553913527 -0.232007145  -0.6571838
#> 797   0.494201101  0.420237732  -9.3034516
#> 798  -0.199593915  0.071890926  10.1017754
#> 799   0.187593927  0.047413466   4.1068136
#> 800  -0.367515081 -0.416628475 -24.7481263
#> 801   0.110652227 -0.041694042  18.6402481
#> 802  -0.042126342  0.240426589 -15.8083025
#> 803   0.015160589  0.096498564  11.1708254
#> 804   0.259617182  0.030087643   6.5784242
#> 805  -0.114958113 -0.029718366 -19.5840896
#> 806   0.576386269 -0.429798182   6.8257190
#> 807  -0.602683375  0.228882797  19.3698408
#> 808  -0.225999864  0.311782170 -17.0859896
#> 809   0.500639373 -0.145613453   2.9720687
#> 810  -0.209494722 -0.426561716  -6.3184209
#> 811   0.266374238  0.085744146  27.1245118
#> 812  -0.410138928  0.259405558 -15.2716605
#> 813   0.279150234  0.110883493  -0.2339727
#> 814  -0.290107507 -0.410312574   1.3784473
#> 815   0.194006044  0.043373860  13.8365258
#> 816   0.287270225  0.006006207 -25.5056162
#> 817  -0.247345772  0.345498632  23.1938603
#> 818  -0.223125646  0.091222239  -3.7165462
#> 819   0.308227185 -0.324993611 -14.6373280
#> 820  -0.048358525 -0.120235852  -4.4740984
#> 821  -0.156614736  0.169464389  16.3206098
#> 822  -0.201417377  0.247640457  -5.6548240
#> 823   0.227071468 -0.401529667   2.2748754
#> 824  -0.007555453  0.439635987  -0.1595699
#> 825   0.076316690 -0.314473108 -16.7198606
#> 826  -0.154026208 -0.225840885  22.9991098
#> 827   0.168597472  0.190423758  -5.9523314
#> 828   0.052980017 -0.084926730  -2.0215337
#> 829   0.099078114 -0.001768670  -8.9893929
#> 830  -0.343826751 -0.102519647  -5.0617928
#> 831   0.098866692  0.462831515  12.4706444
#> 832  -0.098829186 -0.178553024  15.9069312
#> 833   0.060632189 -0.176701408 -15.6250145
#> 834   0.381343433  0.296983995  -3.1182104
#> 835  -0.072160660 -0.031191856  -2.6539223
#> 836   0.055006456  0.121094807 -14.9181948
#> 837  -0.052987426 -0.232109650  10.3043758
#> 838  -0.087344022 -0.139010352   5.2136938
#> 839  -0.464659827  0.275069285  -5.6574819
#> 840   0.276124135  0.183689660  11.2296970
#> 841   0.252190060 -0.334063632 -14.0395285
#> 842  -0.498715175  0.211003463  -1.4726871
#> 843   0.196162356 -0.106494373  22.5695662
#> 844   0.365816733  0.023501566 -18.1036941
#> 845  -0.327335562 -0.253272418  14.2016053
#> 846   0.469663200 -0.129367779 -11.0213202
#> 847  -0.380048363  0.036333957   2.1232694
#> 848   0.171805546 -0.030496767   6.1191398
#> 849  -0.115372180  0.174253518   5.9530589
#> 850  -0.108892981  0.008965954   5.8287910
#> 851   0.148398072  0.344087349  -8.4141645
#> 852  -0.393941807 -0.077807151  -7.0315408
#> 853   0.170514924 -0.390017424  10.1251790
#> 854   0.238443951  0.145376397   9.7358719
#> 855  -0.094535819 -0.056639732 -20.1823975
#> 856   0.065881376  0.374713737   2.2348351
#> 857   0.234884927 -0.471735681   5.8700792
#> 858  -0.148110250  0.143750196 -16.1936983
#> 859  -0.308178899 -0.075366487  -2.8807672
#> 860   0.188572057  0.195538782  15.6524780
#> 861  -0.020740220 -0.200544173   0.9213956
#> 862  -0.332482690  0.466449393  -3.3260430
#> 863   0.497876617 -0.220982158  10.3977071
#> 864   0.310308933 -0.053973714   3.3022905
#> 865  -0.438139663 -0.123242493  -4.6348181
#> 866   0.052886841  0.037157367 -15.2681432
#> 867  -0.360731318  0.193245697   5.8782861
#> 868   0.141844932 -0.218080487  18.3319071
#> 869   0.168120615  0.164819993 -15.3172639
#> 870  -0.150499798  0.081963290   9.1542837
#> 871   0.024583441 -0.321352802 -15.2441627
#> 872   0.234528014 -0.120958917  -5.2507718
#> 873  -0.268394621  0.395212915  10.8397037
#> 874   0.271894354 -0.028963579 -32.2443997
#> 875  -0.216821832 -0.066195433  20.4521392
#> 876   0.311985574 -0.111666356  20.2597409
#> 877  -0.335171869  0.040564023 -11.2149940
#> 878  -0.346956004  0.338705125   9.8190935
#> 879   0.328535109 -0.114462901 -20.9018006
#> 880   0.556839345 -0.136797758   8.4664328
#> 881  -0.419639128  0.013890552   9.6341730
#> 882  -0.064524237  0.099644461   8.8114199
#> 883   0.176926501  0.113715105 -15.9523425
#> 884   0.075708146 -0.246235811   1.8527718
#> 885  -0.890163772  0.018317135   1.7636756
#> 886   0.930205362  0.122954804  15.3722437
#> 887  -0.165200263 -0.200741498 -11.7759961
#> 888   0.215633949  0.120578657  -2.8702685
#> 889  -0.210137306 -0.092766825  11.5393091
#> 890  -0.066097555 -0.182021248  -1.2980116
#> 891  -0.287029791 -0.081243060 -15.8531082
#> 892   0.096445113  0.444914434   4.7521878
#> 893   0.413309252  0.040419813  15.3440911
#> 894   0.197335351 -0.463260634 -18.3280455
#> 895  -0.359951846  0.248674147   3.3711593
#> 896  -0.553913527 -0.232007145  -0.6571838
#> 897   0.494201101  0.420237732  -9.3034516
#> 898  -0.199593915  0.071890926  10.1017754
#> 899   0.187593927  0.047413466   4.1068136
#> 900  -0.367515081 -0.416628475 -24.7481263
#> 901   0.110652227 -0.041694042  18.6402481
#> 902  -0.042126342  0.240426589 -15.8083025
#> 903   0.015160589  0.096498564  11.1708254
#> 904   0.259617182  0.030087643   6.5784242
#> 905  -0.114958113 -0.029718366 -19.5840896
#> 906   0.576386269 -0.429798182   6.8257190
#> 907  -0.602683375  0.228882797  19.3698408
#> 908  -0.225999864  0.311782170 -17.0859896
#> 909   0.500639373 -0.145613453   2.9720687
#> 910  -0.209494722 -0.426561716  -6.3184209
#> 911   0.266374238  0.085744146  27.1245118
#> 912  -0.410138928  0.259405558 -15.2716605
#> 913   0.279150234  0.110883493  -0.2339727
#> 914  -0.290107507 -0.410312574   1.3784473
#> 915   0.194006044  0.043373860  13.8365258
#> 916   0.287270225  0.006006207 -25.5056162
#> 917  -0.247345772  0.345498632  23.1938603
#> 918  -0.223125646  0.091222239  -3.7165462
#> 919   0.308227185 -0.324993611 -14.6373280
#> 920  -0.048358525 -0.120235852  -4.4740984
#> 921  -0.156614736  0.169464389  16.3206098
#> 922  -0.201417377  0.247640457  -5.6548240
#> 923   0.227071468 -0.401529667   2.2748754
#> 924  -0.007555453  0.439635987  -0.1595699
#> 925   0.076316690 -0.314473108 -16.7198606
#> 926  -0.154026208 -0.225840885  22.9991098
#> 927   0.168597472  0.190423758  -5.9523314
#> 928   0.052980017 -0.084926730  -2.0215337
#> 929   0.099078114 -0.001768670  -8.9893929
#> 930  -0.343826751 -0.102519647  -5.0617928
#> 931   0.098866692  0.462831515  12.4706444
#> 932  -0.098829186 -0.178553024  15.9069312
#> 933   0.060632189 -0.176701408 -15.6250145
#> 934   0.381343433  0.296983995  -3.1182104
#> 935  -0.072160660 -0.031191856  -2.6539223
#> 936   0.055006456  0.121094807 -14.9181948
#> 937  -0.052987426 -0.232109650  10.3043758
#> 938  -0.087344022 -0.139010352   5.2136938
#> 939  -0.464659827  0.275069285  -5.6574819
#> 940   0.276124135  0.183689660  11.2296970
#> 941   0.252190060 -0.334063632 -14.0395285
#> 942  -0.498715175  0.211003463  -1.4726871
#> 943   0.196162356 -0.106494373  22.5695662
#> 944   0.365816733  0.023501566 -18.1036941
#> 945  -0.327335562 -0.253272418  14.2016053
#> 946   0.469663200 -0.129367779 -11.0213202
#> 947  -0.380048363  0.036333957   2.1232694
#> 948   0.171805546 -0.030496767   6.1191398
#> 949  -0.115372180  0.174253518   5.9530589
#> 950  -0.108892981  0.008965954   5.8287910
#> 951   0.148398072  0.344087349  -8.4141645
#> 952  -0.393941807 -0.077807151  -7.0315408
#> 953   0.170514924 -0.390017424  10.1251790
#> 954   0.238443951  0.145376397   9.7358719
#> 955  -0.094535819 -0.056639732 -20.1823975
#> 956   0.065881376  0.374713737   2.2348351
#> 957   0.234884927 -0.471735681   5.8700792
#> 958  -0.148110250  0.143750196 -16.1936983
#> 959  -0.308178899 -0.075366487  -2.8807672
#> 960   0.188572057  0.195538782  15.6524780
#> 961  -0.020740220 -0.200544173   0.9213956
#> 962  -0.332482690  0.466449393  -3.3260430
#> 963   0.497876617 -0.220982158  10.3977071
#> 964   0.310308933 -0.053973714   3.3022905
#> 965  -0.438139663 -0.123242493  -4.6348181
#> 966   0.052886841  0.037157367 -15.2681432
#> 967  -0.360731318  0.193245697   5.8782861
#> 968   0.141844932 -0.218080487  18.3319071
#> 969   0.168120615  0.164819993 -15.3172639
#> 970  -0.150499798  0.081963290   9.1542837
#> 971   0.024583441 -0.321352802 -15.2441627
#> 972   0.234528014 -0.120958917  -5.2507718
#> 973  -0.268394621  0.395212915  10.8397037
#> 974   0.271894354 -0.028963579 -32.2443997
#> 975  -0.216821832 -0.066195433  20.4521392
#> 976   0.311985574 -0.111666356  20.2597409
#> 977  -0.335171869  0.040564023 -11.2149940
#> 978  -0.346956004  0.338705125   9.8190935
#> 979   0.328535109 -0.114462901 -20.9018006
#> 980   0.556839345 -0.136797758   8.4664328
#> 981  -0.419639128  0.013890552   9.6341730
#> 982  -0.064524237  0.099644461   8.8114199
#> 983   0.176926501  0.113715105 -15.9523425
#> 984   0.075708146 -0.246235811   1.8527718
#> 985  -0.890163772  0.018317135   1.7636756
#> 986   0.930205362  0.122954804  15.3722437
#> 987  -0.165200263 -0.200741498 -11.7759961
#> 988   0.215633949  0.120578657  -2.8702685
#> 989  -0.210137306 -0.092766825  11.5393091
#> 990  -0.066097555 -0.182021248  -1.2980116
#> 991  -0.287029791 -0.081243060 -15.8531082
#> 992   0.096445113  0.444914434   4.7521878
#> 993   0.413309252  0.040419813  15.3440911
#> 994   0.197335351 -0.463260634 -18.3280455
#> 995  -0.359951846  0.248674147   3.3711593
#> 996  -0.553913527 -0.232007145  -0.6571838
#> 997   0.494201101  0.420237732  -9.3034516
#> 998  -0.199593915  0.071890926  10.1017754
#> 999   0.187593927  0.047413466   4.1068136
#> 1000 -0.367515081 -0.416628475 -24.7481263

rowdiff(yields[1:5,], exclude="factor", na.rm = TRUE, 
na_action = "get_mode",direction = "reverse")

#>        height      weight      yield
#> 1 -0.04212634  0.24042659 -15.808303
#> 2 -0.04212634  0.24042659 -15.808303
#> 3  0.01516059  0.09649856  11.170825
#> 4  0.25961718  0.03008764   6.578424
#> 5 -0.11495811 -0.02971837 -19.584090