Get row differences between values

This function returns the differences between rows depending on the user's choice.

rowdiff(
  df,
  direction = "forward",
  exclude = NULL,
  na.rm = FALSE,
  na_action = NULL,
  ...
)

Arguments

df
The data set for which differences are required

direction
One of forward and reverse. The default is forward meaning the differences are calculated in such a way that the difference between the current value and the next is returned

exclude
A character vector specifying what classes should be removed. See examples below

na.rm
Logical. Should missing values be removed? The missing values referred to are those introduced during the calculation ie when subtracting a row with itself. Defaults to FALSE.

na_action
If na.rm is TRUE, how should missing values be replaced? Depending on the value as set out in `na_replace`, the value can be replaced as per the user's requirement.

...
Other arguments to `na_replace`.

Value

A data.frame object of row differences

See also

Examples

# Remove factor columns
rowdiff(iris,exclude = "factor",direction = "reverse")

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1             NA          NA           NA          NA
#> 2           -0.2        -0.5          0.0         0.0
#> 3           -0.2         0.2         -0.1         0.0
#> 4           -0.1        -0.1          0.2         0.0
#> 5            0.4         0.5         -0.1         0.0
#> 6            0.4         0.3          0.3         0.2
#> 7           -0.8        -0.5         -0.3        -0.1
#> 8            0.4         0.0          0.1        -0.1
#> 9           -0.6        -0.5         -0.1         0.0
#> 10           0.5         0.2          0.1        -0.1
#> 11           0.5         0.6          0.0         0.1
#> 12          -0.6        -0.3          0.1         0.0
#> 13           0.0        -0.4         -0.2        -0.1
#> 14          -0.5         0.0         -0.3         0.0
#> 15           1.5         1.0          0.1         0.1
#> 16          -0.1         0.4          0.3         0.2
#> 17          -0.3        -0.5         -0.2         0.0
#> 18          -0.3        -0.4          0.1        -0.1
#> 19           0.6         0.3          0.3         0.0
#> 20          -0.6         0.0         -0.2         0.0
#> 21           0.3        -0.4          0.2        -0.1
#> 22          -0.3         0.3         -0.2         0.2
#> 23          -0.5        -0.1         -0.5        -0.2
#> 24           0.5        -0.3          0.7         0.3
#> 25          -0.3         0.1          0.2        -0.3
#> 26           0.2        -0.4         -0.3         0.0
#> 27           0.0         0.4          0.0         0.2
#> 28           0.2         0.1         -0.1        -0.2
#> 29           0.0        -0.1         -0.1         0.0
#> 30          -0.5        -0.2          0.2         0.0
#> 31           0.1        -0.1          0.0         0.0
#> 32           0.6         0.3         -0.1         0.2
#> 33          -0.2         0.7          0.0        -0.3
#> 34           0.3         0.1         -0.1         0.1
#> 35          -0.6        -1.1          0.1         0.0
#> 36           0.1         0.1         -0.3         0.0
#> 37           0.5         0.3          0.1         0.0
#> 38          -0.6         0.1          0.1        -0.1
#> 39          -0.5        -0.6         -0.1         0.1
#> 40           0.7         0.4          0.2         0.0
#> 41          -0.1         0.1         -0.2         0.1
#> 42          -0.5        -1.2          0.0         0.0
#> 43          -0.1         0.9          0.0        -0.1
#> 44           0.6         0.3          0.3         0.4
#> 45           0.1         0.3          0.3        -0.2
#> 46          -0.3        -0.8         -0.5        -0.1
#> 47           0.3         0.8          0.2        -0.1
#> 48          -0.5        -0.6         -0.2         0.0
#> 49           0.7         0.5          0.1         0.0
#> 50          -0.3        -0.4         -0.1         0.0
#> 51           2.0        -0.1          3.3         1.2
#> 52          -0.6         0.0         -0.2         0.1
#> 53           0.5        -0.1          0.4         0.0
#> 54          -1.4        -0.8         -0.9        -0.2
#> 55           1.0         0.5          0.6         0.2
#> 56          -0.8         0.0         -0.1        -0.2
#> 57           0.6         0.5          0.2         0.3
#> 58          -1.4        -0.9         -1.4        -0.6
#> 59           1.7         0.5          1.3         0.3
#> 60          -1.4        -0.2         -0.7         0.1
#> 61          -0.2        -0.7         -0.4        -0.4
#> 62           0.9         1.0          0.7         0.5
#> 63           0.1        -0.8         -0.2        -0.5
#> 64           0.1         0.7          0.7         0.4
#> 65          -0.5         0.0         -1.1        -0.1
#> 66           1.1         0.2          0.8         0.1
#> 67          -1.1        -0.1          0.1         0.1
#> 68           0.2        -0.3         -0.4        -0.5
#> 69           0.4        -0.5          0.4         0.5
#> 70          -0.6         0.3         -0.6        -0.4
#> 71           0.3         0.7          0.9         0.7
#> 72           0.2        -0.4         -0.8        -0.5
#> 73           0.2        -0.3          0.9         0.2
#> 74          -0.2         0.3         -0.2        -0.3
#> 75           0.3         0.1         -0.4         0.1
#> 76           0.2         0.1          0.1         0.1
#> 77           0.2        -0.2          0.4         0.0
#> 78          -0.1         0.2          0.2         0.3
#> 79          -0.7        -0.1         -0.5        -0.2
#> 80          -0.3        -0.3         -1.0        -0.5
#> 81          -0.2        -0.2          0.3         0.1
#> 82           0.0         0.0         -0.1        -0.1
#> 83           0.3         0.3          0.2         0.2
#> 84           0.2         0.0          1.2         0.4
#> 85          -0.6         0.3         -0.6        -0.1
#> 86           0.6         0.4          0.0         0.1
#> 87           0.7        -0.3          0.2        -0.1
#> 88          -0.4        -0.8         -0.3        -0.2
#> 89          -0.7         0.7         -0.3         0.0
#> 90          -0.1        -0.5         -0.1         0.0
#> 91           0.0         0.1          0.4        -0.1
#> 92           0.6         0.4          0.2         0.2
#> 93          -0.3        -0.4         -0.6        -0.2
#> 94          -0.8        -0.3         -0.7        -0.2
#> 95           0.6         0.4          0.9         0.3
#> 96           0.1         0.3          0.0        -0.1
#> 97           0.0        -0.1          0.0         0.1
#> 98           0.5         0.0          0.1         0.0
#> 99          -1.1        -0.4         -1.3        -0.2
#> 100          0.6         0.3          1.1         0.2
#> 101          0.6         0.5          1.9         1.2
#> 102         -0.5        -0.6         -0.9        -0.6
#> 103          1.3         0.3          0.8         0.2
#> 104         -0.8        -0.1         -0.3        -0.3
#> 105          0.2         0.1          0.2         0.4
#> 106          1.1         0.0          0.8        -0.1
#> 107         -2.7        -0.5         -2.1        -0.4
#> 108          2.4         0.4          1.8         0.1
#> 109         -0.6        -0.4         -0.5         0.0
#> 110          0.5         1.1          0.3         0.7
#> 111         -0.7        -0.4         -1.0        -0.5
#> 112         -0.1        -0.5          0.2        -0.1
#> 113          0.4         0.3          0.2         0.2
#> 114         -1.1        -0.5         -0.5        -0.1
#> 115          0.1         0.3          0.1         0.4
#> 116          0.6         0.4          0.2        -0.1
#> 117          0.1        -0.2          0.2        -0.5
#> 118          1.2         0.8          1.2         0.4
#> 119          0.0        -1.2          0.2         0.1
#> 120         -1.7        -0.4         -1.9        -0.8
#> 121          0.9         1.0          0.7         0.8
#> 122         -1.3        -0.4         -0.8        -0.3
#> 123          2.1         0.0          1.8         0.0
#> 124         -1.4        -0.1         -1.8        -0.2
#> 125          0.4         0.6          0.8         0.3
#> 126          0.5        -0.1          0.3        -0.3
#> 127         -1.0        -0.4         -1.2         0.0
#> 128         -0.1         0.2          0.1         0.0
#> 129          0.3        -0.2          0.7         0.3
#> 130          0.8         0.2          0.2        -0.5
#> 131          0.2        -0.2          0.3         0.3
#> 132          0.5         1.0          0.3         0.1
#> 133         -1.5        -1.0         -0.8         0.2
#> 134         -0.1         0.0         -0.5        -0.7
#> 135         -0.2        -0.2          0.5        -0.1
#> 136          1.6         0.4          0.5         0.9
#> 137         -1.4         0.4         -0.5         0.1
#> 138          0.1        -0.3         -0.1        -0.6
#> 139         -0.4        -0.1         -0.7         0.0
#> 140          0.9         0.1          0.6         0.3
#> 141         -0.2         0.0          0.2         0.3
#> 142          0.2         0.0         -0.5        -0.1
#> 143         -1.1        -0.4          0.0        -0.4
#> 144          1.0         0.5          0.8         0.4
#> 145         -0.1         0.1         -0.2         0.2
#> 146          0.0        -0.3         -0.5        -0.2
#> 147         -0.4        -0.5         -0.2        -0.4
#> 148          0.2         0.5          0.2         0.1
#> 149         -0.3         0.4          0.2         0.3
#> 150         -0.3        -0.4         -0.3        -0.5

rowdiff(iris[1:5,], exclude="factor", na.rm = TRUE, na_action = "get_mode",direction = "reverse")

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1         -0.2        -0.5          0.0           0
#> 2         -0.2        -0.5          0.0           0
#> 3         -0.2         0.2         -0.1           0
#> 4         -0.1        -0.1          0.2           0
#> 5          0.4         0.5         -0.1           0