A convenient selector gadget

A convenient selector gadget

select_col(df, ...)

Arguments

df
The data set from which to select a column

...
columns to select, no quotes

Value

Returns a dataframe with selected columns

Details

A friendly way to select a column or several columns. Mainly for non-pipe usage It is recommended to use known select functions to do pipe manipulations. Otherwise convert to tibble

Examples

select_col(iris,Petal.Length,Sepal.Length,Species,Petal.Width)

#>     Petal.Length Sepal.Length    Species Petal.Width
#> 1            1.4          5.1     setosa         0.2
#> 2            1.4          4.9     setosa         0.2
#> 3            1.3          4.7     setosa         0.2
#> 4            1.5          4.6     setosa         0.2
#> 5            1.4          5.0     setosa         0.2
#> 6            1.7          5.4     setosa         0.4
#> 7            1.4          4.6     setosa         0.3
#> 8            1.5          5.0     setosa         0.2
#> 9            1.4          4.4     setosa         0.2
#> 10           1.5          4.9     setosa         0.1
#> 11           1.5          5.4     setosa         0.2
#> 12           1.6          4.8     setosa         0.2
#> 13           1.4          4.8     setosa         0.1
#> 14           1.1          4.3     setosa         0.1
#> 15           1.2          5.8     setosa         0.2
#> 16           1.5          5.7     setosa         0.4
#> 17           1.3          5.4     setosa         0.4
#> 18           1.4          5.1     setosa         0.3
#> 19           1.7          5.7     setosa         0.3
#> 20           1.5          5.1     setosa         0.3
#> 21           1.7          5.4     setosa         0.2
#> 22           1.5          5.1     setosa         0.4
#> 23           1.0          4.6     setosa         0.2
#> 24           1.7          5.1     setosa         0.5
#> 25           1.9          4.8     setosa         0.2
#> 26           1.6          5.0     setosa         0.2
#> 27           1.6          5.0     setosa         0.4
#> 28           1.5          5.2     setosa         0.2
#> 29           1.4          5.2     setosa         0.2
#> 30           1.6          4.7     setosa         0.2
#> 31           1.6          4.8     setosa         0.2
#> 32           1.5          5.4     setosa         0.4
#> 33           1.5          5.2     setosa         0.1
#> 34           1.4          5.5     setosa         0.2
#> 35           1.5          4.9     setosa         0.2
#> 36           1.2          5.0     setosa         0.2
#> 37           1.3          5.5     setosa         0.2
#> 38           1.4          4.9     setosa         0.1
#> 39           1.3          4.4     setosa         0.2
#> 40           1.5          5.1     setosa         0.2
#> 41           1.3          5.0     setosa         0.3
#> 42           1.3          4.5     setosa         0.3
#> 43           1.3          4.4     setosa         0.2
#> 44           1.6          5.0     setosa         0.6
#> 45           1.9          5.1     setosa         0.4
#> 46           1.4          4.8     setosa         0.3
#> 47           1.6          5.1     setosa         0.2
#> 48           1.4          4.6     setosa         0.2
#> 49           1.5          5.3     setosa         0.2
#> 50           1.4          5.0     setosa         0.2
#> 51           4.7          7.0 versicolor         1.4
#> 52           4.5          6.4 versicolor         1.5
#> 53           4.9          6.9 versicolor         1.5
#> 54           4.0          5.5 versicolor         1.3
#> 55           4.6          6.5 versicolor         1.5
#> 56           4.5          5.7 versicolor         1.3
#> 57           4.7          6.3 versicolor         1.6
#> 58           3.3          4.9 versicolor         1.0
#> 59           4.6          6.6 versicolor         1.3
#> 60           3.9          5.2 versicolor         1.4
#> 61           3.5          5.0 versicolor         1.0
#> 62           4.2          5.9 versicolor         1.5
#> 63           4.0          6.0 versicolor         1.0
#> 64           4.7          6.1 versicolor         1.4
#> 65           3.6          5.6 versicolor         1.3
#> 66           4.4          6.7 versicolor         1.4
#> 67           4.5          5.6 versicolor         1.5
#> 68           4.1          5.8 versicolor         1.0
#> 69           4.5          6.2 versicolor         1.5
#> 70           3.9          5.6 versicolor         1.1
#> 71           4.8          5.9 versicolor         1.8
#> 72           4.0          6.1 versicolor         1.3
#> 73           4.9          6.3 versicolor         1.5
#> 74           4.7          6.1 versicolor         1.2
#> 75           4.3          6.4 versicolor         1.3
#> 76           4.4          6.6 versicolor         1.4
#> 77           4.8          6.8 versicolor         1.4
#> 78           5.0          6.7 versicolor         1.7
#> 79           4.5          6.0 versicolor         1.5
#> 80           3.5          5.7 versicolor         1.0
#> 81           3.8          5.5 versicolor         1.1
#> 82           3.7          5.5 versicolor         1.0
#> 83           3.9          5.8 versicolor         1.2
#> 84           5.1          6.0 versicolor         1.6
#> 85           4.5          5.4 versicolor         1.5
#> 86           4.5          6.0 versicolor         1.6
#> 87           4.7          6.7 versicolor         1.5
#> 88           4.4          6.3 versicolor         1.3
#> 89           4.1          5.6 versicolor         1.3
#> 90           4.0          5.5 versicolor         1.3
#> 91           4.4          5.5 versicolor         1.2
#> 92           4.6          6.1 versicolor         1.4
#> 93           4.0          5.8 versicolor         1.2
#> 94           3.3          5.0 versicolor         1.0
#> 95           4.2          5.6 versicolor         1.3
#> 96           4.2          5.7 versicolor         1.2
#> 97           4.2          5.7 versicolor         1.3
#> 98           4.3          6.2 versicolor         1.3
#> 99           3.0          5.1 versicolor         1.1
#> 100          4.1          5.7 versicolor         1.3
#> 101          6.0          6.3  virginica         2.5
#> 102          5.1          5.8  virginica         1.9
#> 103          5.9          7.1  virginica         2.1
#> 104          5.6          6.3  virginica         1.8
#> 105          5.8          6.5  virginica         2.2
#> 106          6.6          7.6  virginica         2.1
#> 107          4.5          4.9  virginica         1.7
#> 108          6.3          7.3  virginica         1.8
#> 109          5.8          6.7  virginica         1.8
#> 110          6.1          7.2  virginica         2.5
#> 111          5.1          6.5  virginica         2.0
#> 112          5.3          6.4  virginica         1.9
#> 113          5.5          6.8  virginica         2.1
#> 114          5.0          5.7  virginica         2.0
#> 115          5.1          5.8  virginica         2.4
#> 116          5.3          6.4  virginica         2.3
#> 117          5.5          6.5  virginica         1.8
#> 118          6.7          7.7  virginica         2.2
#> 119          6.9          7.7  virginica         2.3
#> 120          5.0          6.0  virginica         1.5
#> 121          5.7          6.9  virginica         2.3
#> 122          4.9          5.6  virginica         2.0
#> 123          6.7          7.7  virginica         2.0
#> 124          4.9          6.3  virginica         1.8
#> 125          5.7          6.7  virginica         2.1
#> 126          6.0          7.2  virginica         1.8
#> 127          4.8          6.2  virginica         1.8
#> 128          4.9          6.1  virginica         1.8
#> 129          5.6          6.4  virginica         2.1
#> 130          5.8          7.2  virginica         1.6
#> 131          6.1          7.4  virginica         1.9
#> 132          6.4          7.9  virginica         2.0
#> 133          5.6          6.4  virginica         2.2
#> 134          5.1          6.3  virginica         1.5
#> 135          5.6          6.1  virginica         1.4
#> 136          6.1          7.7  virginica         2.3
#> 137          5.6          6.3  virginica         2.4
#> 138          5.5          6.4  virginica         1.8
#> 139          4.8          6.0  virginica         1.8
#> 140          5.4          6.9  virginica         2.1
#> 141          5.6          6.7  virginica         2.4
#> 142          5.1          6.9  virginica         2.3
#> 143          5.1          5.8  virginica         1.9
#> 144          5.9          6.8  virginica         2.3
#> 145          5.7          6.7  virginica         2.5
#> 146          5.2          6.7  virginica         2.3
#> 147          5.0          6.3  virginica         1.9
#> 148          5.2          6.5  virginica         2.0
#> 149          5.4          6.2  virginica         2.3
#> 150          5.1          5.9  virginica         1.8

# A pipe friendly example
if (FALSE) {
library(dplyr)
as_tibble(iris) %>%
select_col(Species)
}